==> code/03_alpha_diversity.Rmd <==
---
title: "Alpha diversity"
author: "Haipeng Sun"
date: "7/14/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(phyloseq)
library(purrr)
library(furrr)
library(picante)
library(vegan)
library(MASS)
library(ggpubr)
library(rstatix)
library(tidyverse)
```

## load data
```{r}
path_rec = "~/Dropbox/Rutgers/20220722_IFNH_CSR_new/"
theme_custom =  theme_bw() + theme(panel.background = element_rect(fill = NA), 
                                   text = element_text(face = 'bold', color = "black"),
                                   axis.text = element_text(size = 12, color = "black"),
                                   axis.title = element_text(size = 14),
                                   legend.text = element_text(size= 14),
                                   legend.title = element_text(size= 14))
load(file = paste0(path_rec,"data/general_data.Rdata"))
load(file = paste0(path_rec,"data/BW_responder.RData"))
load(file = paste0(path_rec,"data/alpha.Rdata"))
dat_alpha <- left_join(alpha_p, meta_all, by = "SampleID")

linear_model <- function(data, x, y) {
  #input data, respond variable y, variable formula
  require(effectsize)
  fit = lm(as.formula(paste0(y, " ~ ", x)), data = data)
  anova_fit <- anova(fit)
  sum_fit <- summary(fit)
  effect_fit <- effectsize(anova_fit, "omega")
  res_coef <- sum_fit$coefficients %>% as.data.frame() %>% 
    rownames_to_column(var="variable") %>% mutate(Metrics = y)
  res_anova <- right_join(anova_fit %>% as.data.frame() %>% rownames_to_column(var = "Parameter"),
                          effect_fit %>% as.data.frame(), by = "Parameter") %>% 
    mutate(Metrics = y)
  res <- list(res_coef, res_anova)
  names(res) = c("coef","anova")
  res  
}

comp_kw_aov = function(wdat, x, y){
  #input data, respond variable y, categorize variable x
  require(agricolae)
  ktest1 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="none", group=T, main = "wdat", console=FALSE)
  out_g_1 = ktest1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(KW_group = groups)
  ktest2 = kruskal(wdat[y], wdat[x], alpha = 0.05, p.adj="none", group=F, main = "wdat", console=FALSE)
  out_c_1 = ktest2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(KW_pvalue = pvalue)
  fit = aov(as.formula(paste0(y, "~", x)), wdat)
  tukey1 = HSD.test(fit, x, group = T)
  out_g_2 = tukey1$groups %>% rownames_to_column(var = "Var") %>% dplyr::select(Var, groups) %>% rename(aov_group = groups)
  tukey2 = HSD.test(fit, x, group = F)
  out_c_2 = tukey2$comparison %>% rownames_to_column(var = "comp") %>% dplyr::select(comp, pvalue) %>% rename(aov_pvalue = pvalue)
  out_g = merge(out_g_1, out_g_2,by = 1)
  out_c = merge(out_c_1, out_c_2,by = 1)
  res = list(out_g, out_c)
  names(res) = c("grp","comp")
  res
}
```


## alpha expolore mom vaginal
```{r}
wdat <- dat_alpha %>% mutate(grp = case_when(TimePoint == "Week3" ~ "Week3",
                                             TimePoint == "Week6" ~ "Week6",
                                             TimePoint == "Week9-14" ~ "Week9-14",
                                             TimePoint == "Week18" ~ "Week18",
                                             BodySite == "Vaginal" ~ "At birth")) %>%
  mutate(grp = factor(grp, levels = c("At birth","Week3","Week6","Week9-14","Week18"),
                      ordered = TRUE)) 

# my_comparisons <- wdat$grp %>% unique() %>% as.character() %>% combn(., 2) %>% apply(.,2, list)
# comp_comb <- wdat$grp %>% unique() %>% as.character() %>% combn(., 2)
# my_comparisons <- lapply(1:ncol(comp_comb),function(x) comp_comb[,x])

wwdat <- wdat %>% dplyr::select(SampleID,grp,BodySite,Faith_PD:Pielou_evenness) %>% 
             pivot_longer(c(Faith_PD:Pielou_evenness), names_to = "Metrics")

stat.test <- wwdat %>% group_by(Metrics) %>% 
  rstatix::wilcox_test(value ~ grp) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "grp", dodge = 0.8, scales = "free_y") %>% 
  filter(!p.adj.signif=="ns") 

g <-ggplot(wwdat, aes(x=grp, y=value)) + 
  geom_boxplot(aes(color=BodySite, fill=BodySite), alpha = 0.3, outlier.alpha = 0) + 
  geom_point(aes(color=BodySite), position = position_dodge2(width = 0.2)) + 
  stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0, size = 12/.pt, hide.ns = T) +
  #stat_compare_means(comparisons = my_comparisons, label = "p.signif", tip.length = 0, hide.ns = T) +
  scale_color_manual(values = c("Feces"="#7fc97f","Vaginal"="#beaed4"), 
                     name = "",
                     labels = c("pups fecal", "moms vaginal"),
                     aesthetics = c("color","fill")) + 
  labs(x="",y="", tittle="") + 
  facet_wrap(~ Metrics, nrow = 2, scales = "free_y") + 
  theme_custom + theme(aspect.ratio = 0.8,
                       panel.background = element_rect(fill = NA),
                       strip.text = element_text(size=12,color="black",face='bold'),
                       axis.title = element_text(size=12,color="black",face='bold'),
                       axis.text = element_text(size=12,color="black",face='bold'),
                       axis.text.x = element_text(size=12,color="black",face='bold',angle=90),
                       axis.text.y = element_text(size=12,color="black",face='bold'),
                       legend.text = element_text(size=12,color="black",face='bold'),
                       legend.title = element_text(size=12,color="black",face='bold'),
                       title = element_text(size=12,color="black",face='bold'))

ggsave(filename = paste0(temp_out,"alpha/vaginal_fecal_by_Age_group.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
```

## alpha vs DNAconc
```{r}
dat <- dat_alpha %>% filter(!is.na(DNA_Conc)) %>% 
  dplyr::select(SampleID, DNA_Conc, Faith_PD:Pielou_evenness) %>% 
  pivot_longer(Faith_PD:Pielou_evenness, names_to = "Metrics", values_to = "value")

g <- ggplot(dat, aes(DNA_Conc,value)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  stat_regline_equation(label.x.npc = 0.2,label.y.npc = 0.2) + 
  stat_cor(label.x.npc = 0.2, label.y.npc = 0.01) + 
  scale_x_log10() + 
  scale_y_log10() +  
  labs(x=expression("DNA concentration (ng/"*mu*"L)"), y="", title = "") + 
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
                     legend.position="none",
                     panel.background = element_rect(fill = NA),
                     strip.text = element_text(size=12,color="black",face='bold'),
                     axis.title = element_text(size=12,color="black",face='bold'),
                     axis.text = element_text(size=12,color="black",face='bold'),
                     axis.text.x = element_text(size=12,color="black",face='bold',angle=0),
                     axis.text.y = element_text(size=12,color="black",face='bold'),
                     legend.text = element_text(size=12,color="black",face='bold'),
                     legend.title = element_text(size=12,color="black",face='bold'),
                     title = element_text(size=12,color="black",face='bold'))  
ggsave(filename = paste0(temp_out, "alpha/alpha_by_DNAconc.pdf"), device = cairo_pdf, 
         plot = g, width = 9, height = 7, units = "in", dpi = 300) 

dat <- dat_alpha %>% filter(!is.na(PCR_Conc)) %>% 
  dplyr::select(SampleID, PCR_Conc, Faith_PD:Pielou_evenness) %>% 
  pivot_longer(Faith_PD:Pielou_evenness, names_to = "Metrics", values_to = "value")

g <- ggplot(dat, aes(PCR_Conc,value)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  stat_regline_equation(label.x.npc = 0.2,label.y.npc = 0.2) + 
  stat_cor(label.x.npc = 0.2, label.y.npc = 0.01) + 
  scale_x_log10() + 
  scale_y_log10() +  
  labs(x=expression("PCR concentration (ng/"*mu*"L)"), y="", title = "") + 
  facet_wrap(~Metrics, nrow = 2, scales = "free") +
  theme_bw() + theme(#aspect.ratio = 0.8,
                     legend.position="none",
                     panel.background = element_rect(fill = NA),
                     strip.text = element_text(size=12,color="black",face='bold'),
                     axis.title = element_text(size=12,color="black",face='bold'),
                     axis.text = element_text(size=12,color="black",face='bold'),
                     axis.text.x = element_text(size=12,color="black",face='bold',angle=0),
                     axis.text.y = element_text(size=12,color="black",face='bold'),
                     legend.text = element_text(size=12,color="black",face='bold'),
                     legend.title = element_text(size=12,color="black",face='bold'),
                     title = element_text(size=12,color="black",face='bold')) 
ggsave(filename = paste0(temp_out, "alpha/alpha_by_PCRconc.pdf"), device = cairo_pdf, 
         plot = g, width = 9, height = 7, units = "in", dpi = 300) 

```

## effect of birthmode on pups
### linear model
```{r}
dat_alpha_pup <- dat_alpha %>% filter(Generation=="F1")
metric <- names(alpha_p)[2:5]
###TimePointInWeeks
all_res <- map(metric, linear_model, data = dat_alpha_pup %>% 
                 mutate(TimePointInWeeks = as.character(TimePointInWeeks)), 
               x = "BirthMode + Sex + TimePointInWeeks")
write.csv(rbind(all_res[[1]]$coef,all_res[[2]]$coef,all_res[[3]]$coef,all_res[[4]]$coef),
          file = paste0(temp_out,"alpha/lm_coef_TimePointInWeeks_res.csv"), row.names = FALSE)
write.csv(rbind(all_res[[1]]$anova,all_res[[2]]$anova,all_res[[3]]$anova,all_res[[4]]$anova),
          file = paste0(temp_out,"alpha/lm_anova_TimePointInWeeks_res.csv"), row.names = FALSE)
###Age_group
all_res <-lapply(metric, function(y){
  linear_model(data = dat_alpha_pup, y = y,
                  x = "BirthMode + Sex + TimePoint")
})
write.csv(do.call(rbind,lapply(1:length(all_res), function(n){all_res[[n]]$coef})),
          file = paste0(temp_out,"alpha/lm_coef_Age_group_res.csv"), row.names = FALSE)
write.csv(do.call(rbind,lapply(1:length(all_res), function(n){all_res[[n]]$anova})),
          file = paste0(temp_out,"alpha/lm_anova_Age_group_res.csv"), row.names = FALSE)
```

### compare all stat
```{r}
dat_alpha_pup <- dat_alpha %>% filter(Generation=="F1")
metric <- names(alpha_p)[2:5]
###TimePointInWeeks
lst_df <- expand.grid(dat_alpha_pup$Sex %>% unique(), 
                       dat_alpha_pup$TimePointInWeeks %>% unique(), 
                       metric, stringsAsFactors = F)
names(lst_df) <- c("Sex", "TimePointInWeeks", "Metrics")

all_res <- map_dfr(1:nrow(lst_df), function(n){
  wdat <- dat_alpha_pup %>% filter(Sex == lst_df[n,"Sex"], 
                                  TimePointInWeeks == lst_df[n,"TimePointInWeeks"])
  res <- comp_kw_aov(wdat = wdat, x = "BirthMode", y = lst_df[n,"Metrics"])
  res$grp <- cbind(res$grp, lst_df[n,],row.names = NULL)
  res$comp <- cbind(res$comp, lst_df[n,],row.names = NULL)
  res
})
all_res$comp$KW_padj <- p.adjust(all_res$comp$KW_pvalue, method = "fdr")
all_res$comp$aov_padj <- p.adjust(all_res$comp$aov_pvalue, method = "fdr")
write.csv(all_res$grp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_grp_birthmode_by_Sex_TimePointInWeeks.csv"))
write.csv(all_res$comp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_comp_birthmode_by_Sex_TimePointInWeeks.csv"))

###Age_group
lst_df <- expand.grid(dat_alpha_pup$Sex %>% unique(), 
                       dat_alpha_pup$TimePoint %>% unique(), 
                       metric, stringsAsFactors = F)
names(lst_df) <- c("Sex", "TimePoint", "Metrics")

all_res <- map_dfr(1:nrow(lst_df), function(n){
  wdat <- dat_alpha_pup %>% filter(Sex == lst_df[n,"Sex"], 
                                  TimePoint == lst_df[n,"TimePoint"])
  res <- comp_kw_aov(wdat = wdat, x = "BirthMode", y = lst_df[n,"Metrics"])
  res$grp <- cbind(res$grp, lst_df[n,],row.names = NULL)
  res$comp <- cbind(res$comp, lst_df[n,],row.names = NULL)
  res
})
all_res$comp$KW_padj <- p.adjust(all_res$comp$KW_pvalue, method = "fdr")
all_res$comp$aov_padj <- p.adjust(all_res$comp$aov_pvalue, method = "fdr")
write.csv(all_res$grp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_grp_birthmode_by_Sex_Age_group.csv"))
write.csv(all_res$comp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_comp_birthmode_by_Sex_Age_group.csv"))

```

### plot alpha comparsion
```{r}
library(ggpubr)
dat <- dat_alpha %>% filter(Generation=="F1") %>% 
  dplyr::select(SampleID,Faith_PD,Observed_ASVs,Shannon_Index,Pielou_evenness, 
                BodySite,TimePointInWeeks,TimePoint,Sex, BirthMode) %>% 
  pivot_longer(c(Faith_PD:Pielou_evenness), names_to = "Metrics")

metric <- dat$Metrics %>% unique()

## Age_group + sex
stat.test <- dat %>% group_by(Metrics, TimePoint, Sex) %>% 
  rstatix::wilcox_test(value ~ BirthMode) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "TimePoint", dodge = 0.8, scales = "free_y") %>% 
  filter(!p.adj.signif=="ns") 

for (aa in metric){
  g <- ggplot(data = dat %>% filter(Metrics == aa),
              aes(x = TimePoint, y = value)) + 
    geom_boxplot(aes(color = BirthMode, fill = BirthMode),
                 alpha = 0.3, outlier.alpha = 0) + 
    geom_point(aes(color = BirthMode), alpha = 0.7, size = 1,
               position = position_dodge2(width = 0.6)) + 
    #stat_compare_means(aes(group = BirthMode), hide.ns = T, label =  "p.signif") +
    stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0, size = 12/.pt, hide.ns = T) +
    scale_color_manual(values = BirthMode_color, 
                       name = "Birth Mode",
                       aesthetics = c("color","fill")) + 
    labs(x="",y=aa, tittle="") + 
    facet_wrap(~ Sex, ncol = 2, scales = "free_y", labeller = labeller(Sex = Sex_labs)) + 
    theme_bw() + theme(panel.background = element_rect(fill = NA),
                       strip.text = element_text(size=10,color="black",face='bold'),
                       axis.title = element_text(size=10,color="black",face='bold'),
                       axis.text = element_text(size=10,color="black",face='bold'),
                       axis.text.x = element_text(size=10,color="black",face='bold',angle=0),
                       axis.text.y = element_text(size=10,color="black",face='bold'),
                       legend.text = element_text(size=10,color="black",face='bold'),
                       legend.title = element_text(size=10,color="black",face='bold'),
                       title = element_text(size=10,color="black",face='bold'))
  ggsave(filename = paste0(temp_out,"alpha/pup_birthmode_", aa, "_by_Age_group_sex.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)  
}

#by Sex
stat.test <- dat %>% group_by(Metrics, TimePoint) %>% 
  rstatix::wilcox_test(value ~ Sex) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "TimePoint", dodge = 0.8, scales = "free_y") %>% 
  filter(!p.adj.signif=="ns") 

g <- ggplot(data = dat,
            aes(x = TimePoint, y = value)) + 
  geom_boxplot(aes(color = Sex, fill = Sex),
               alpha = 0.3, outlier.alpha = 0) + 
  geom_point(aes(color = Sex), alpha = 0.7,size = 1,
             position = position_dodge2(width = 0.6)) + 
  #stat_compare_means(aes(group = Sex), hide.ns = T, label =  "p.signif") +
  stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0, size = 12/.pt, hide.ns = T) +
  scale_color_manual(values = c("M"="#1b9e77","F"="#d95f02"), 
                     name = "Sex",
                     labels = c("Male","Female"),
                     aesthetics = c("color","fill")) + 
  labs(x="",y="", tittle="") + 
  facet_wrap(~ Metrics, nrow = 2, scales = "free_y") + 
  theme_bw() + theme(panel.background = element_rect(fill = NA),
                     strip.text = element_text(size=10,color="black",face='bold'),
                     axis.title = element_text(size=10,color="black",face='bold'),
                     axis.text = element_text(size=10,color="black",face='bold'),
                     axis.text.x = element_text(size=10,color="black",face='bold',angle=0),
                     axis.text.y = element_text(size=10,color="black",face='bold'),
                     legend.text = element_text(size=10,color="black",face='bold'),
                     legend.title = element_text(size=10,color="black",face='bold'),
                     title = element_text(size=10,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"alpha/pup_by_sex_by_Age_group.pdf"), device = cairo_pdf, 
       plot = g, width = 7, height = 6, units = "in", dpi = 300)
```

## stat on Age or Sex
```{r}
dat_alpha_pup <- dat_alpha %>% filter(Generation=="F1")
metric <- names(alpha_p)[2:5]

###Sex
lst_df <- expand.grid(dat_alpha_pup$TimePoint %>% unique(), 
                      metric, stringsAsFactors = F)
names(lst_df) <- c("TimePoint", "Metrics")

all_res_2 <- map_dfr(1:nrow(lst_df), function(i){
  wdat <- dat_alpha_pup %>% filter(TimePoint == lst_df[i,"TimePoint"])
  res <- comp_kw_aov(wdat = wdat, x = "Sex", y = lst_df[i,"Metrics"])
  res$grp <- cbind(res$grp, lst_df[i,],row.names = NULL)
  res$comp <- cbind(res$comp, lst_df[i,],row.names = NULL)
  res
})

all_res_2$comp$KW_padj <- p.adjust(all_res_2$comp$KW_pvalue, method = "fdr")
all_res_2$comp$aov_padj <- p.adjust(all_res_2$comp$aov_pvalue, method = "fdr")
write.csv(all_res_2$grp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_grp_Sex_by_Age_group.csv"))
write.csv(all_res_2$comp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_comp_Sex_by_Age_group.csv"))

###Age_group
lst_df <- expand.grid(dat_alpha_pup$Sex %>% unique(), 
                      metric, stringsAsFactors = F)
names(lst_df) <- c("Sex", "Metrics")

all_res_3 <- map(1:nrow(lst_df), function(i){
  wdat <- dat_alpha_pup %>% filter(Sex == lst_df[i,"Sex"])
  res <- comp_kw_aov(wdat = wdat, x = "TimePoint", y = lst_df[i,"Metrics"])
  res$grp <- cbind(res$grp, lst_df[i,],row.names = NULL)
  res$comp <- cbind(res$comp, lst_df[i,],row.names = NULL)
  res
})

dat_grp <- rbind(all_res_3[[1]]$grp,all_res_3[[2]]$grp,all_res_3[[3]]$grp,all_res_3[[4]]$grp, 
                 all_res_3[[5]]$grp, all_res_3[[6]]$grp, all_res_3[[7]]$grp, all_res_3[[8]]$grp)
write.csv(dat_grp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_grp_Age_grp_by_Sex.csv"))

dat_comp <- rbind(all_res_3[[1]]$comp,all_res_3[[2]]$comp,all_res_3[[3]]$comp,all_res_3[[4]]$comp, 
                 all_res_3[[5]]$comp, all_res_3[[6]]$comp, all_res_3[[7]]$comp, all_res_3[[8]]$comp)
dat_comp$KW_padj <- p.adjust(dat_comp$KW_pvalue, method = "fdr")
dat_comp$aov_padj <- p.adjust(dat_comp$aov_pvalue, method = "fdr")

write.csv(dat_comp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_comp_Age_grp_by_Sex.csv"))
```

## overweight 
```{r}
dat <- dat_alpha %>% left_join(at_week15 %>% dplyr::select(MouseID, BW_status, Overweight, Underweight), 
                        by = "MouseID") %>% filter(!is.na(BW_status)) 
metric <- names(alpha_p)[2:5]

all_res <-lapply(metric, function(y){
  linear_model(data = dat, y = y,
                  x = "Overweight + Sex + BirthMode + TimePoint")
})
write.csv(rbind(all_res[[1]]$coef,all_res[[2]]$coef,all_res[[3]]$coef,all_res[[4]]$coef),
          file = paste0(temp_out,"alpha/lm_coef_overweight_res.csv"), row.names = FALSE)
write.csv(rbind(all_res[[1]]$anova,all_res[[2]]$anova,all_res[[3]]$anova,all_res[[4]]$anova),
          file = paste0(temp_out,"alpha/lm_anova_overweight_res.csv"), row.names = FALSE)

lst_df <- expand.grid(dat$Sex %>% unique(),
                      dat$TimePoint %>% unique(), 
                      metric, stringsAsFactors = F)
names(lst_df) <- c("Sex", "TimePoint", "Metrics")

all_res <- map_dfr(1:nrow(lst_df), function(n){
  wdat <- dat %>% filter(TimePoint == lst_df[n,"TimePoint"],
                         Sex == lst_df[n,"Sex"])
  res <- comp_kw_aov(wdat = wdat, x = "Overweight", y = lst_df[n,"Metrics"])
  res$grp <- cbind(res$grp, lst_df[n,],row.names = NULL)
  res$comp <- cbind(res$comp, lst_df[n,],row.names = NULL)
  res
})
all_res$comp$KW_padj <- p.adjust(all_res$comp$KW_pvalue, method = "fdr")
all_res$comp$aov_padj <- p.adjust(all_res$comp$aov_pvalue, method = "fdr")
write.csv(all_res$grp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_grp_overweight_by_TimePointInWeeks.csv"))
write.csv(all_res$comp, row.names = FALSE,
          file = paste0(temp_out,"alpha/kw_aov_comp_overweight_by_TimePointInWeeks.csv"))

dat_long <- dat %>%
  dplyr::select(SampleID,Faith_PD,Observed_ASVs,Shannon_Index,Pielou_evenness, 
         TimePoint, Sex, BirthMode, Overweight) %>% 
  pivot_longer(c(Faith_PD:Pielou_evenness), names_to = "Metrics")

stat.test <- dat_long %>% group_by(Metrics, TimePoint, Sex) %>% 
  rstatix::wilcox_test(value ~ Overweight) %>% 
  rstatix::adjust_pvalue(method = "fdr") %>%
  rstatix::add_significance("p.adj") %>% 
  rstatix::add_xy_position(x = "TimePoint", dodge = 0.8, scales = "free_y") %>% 
  filter(!p.adj.signif=="ns")

for (aa in metric){
  g <- ggplot(dat_long %>% filter(Metrics == aa),
              aes(x = TimePoint, y = value)) + 
    geom_boxplot(aes(color = Overweight, fill = Overweight),
                 alpha = 0.3, outlier.alpha = 0) + 
    geom_point(aes(color = Overweight), alpha = 0.7, size = 1,
               position = position_dodge2(width = 0.9)) +
    #stat_compare_means(aes(group = Overweight), hide.ns = T, label =  "p.signif") +
    stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0, size = 12/.pt, hide.ns = T) +
    scale_color_manual(values = c("Y"="#d73027", "N"="#1a9850"), 
                       name = "Overweight",
                       aesthetics = c("color","fill")) + 
    labs(x="",y="", tittle="") + 
    facet_wrap(~ Sex, ncol = 2, scales = "free_y", labeller = labeller(Sex = Sex_labs)) + 
    theme_bw() + theme(panel.background = element_rect(fill = NA),
                       strip.text = element_text(size=10,color="black",face='bold'),
                       axis.title = element_text(size=10,color="black",face='bold'),
                       axis.text = element_text(size=10,color="black",face='bold'),
                       axis.text.x = element_text(size=10,color="black",face='bold',angle=0),
                       axis.text.y = element_text(size=10,color="black",face='bold'),
                       legend.text = element_text(size=10,color="black",face='bold'),
                       legend.title = element_text(size=10,color="black",face='bold'),
                       title = element_text(size=10,color="black",face='bold'))
  ggsave(filename = paste0(temp_out,"alpha/pup_Overweight_",aa, "_by_Age_group_sex.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 6, units = "in", dpi = 300)
}

```
