==> code/04_beta_diversity.Rmd <==
---
title: "Beta diversity"
author: "Haipeng Sun"
date: "7/14/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(knitr)
library(phyloseq)
library(usedist)
library(purrr)
library(furrr)
library(vegan)
library(ape)
library(rlist)
library(ggpubr)
library(tidyverse)
```

## load data
```{r}
path_rec = "~/Dropbox/Rutgers/20220722_IFNH_CSR_new/"
load(file = paste0(path_rec,"data/general_data.Rdata"))
load(file = paste0(path_rec,"data/BW_responder.RData"))
load(file = paste0(path_rec,"data/beta_DM_uu.RData"))
load(file = paste0(path_rec,"data/beta_DM_wu.RData"))
load(file = paste0(path_rec,"data/beta_DM_bray.RData"))
load(file = paste0(path_rec,"data/beta_DM_jaccard.RData"))

calcOmega2_dm <- function(x) {
    # Computes the effect size omega^2 parital using the output from adonis test
    # Args:
    #   x: adonis output
    # Returns:
    #   A dataframe with calculated effect size omega^2 partial added to original dataframe
    require(dplyr)
    N_t = x["Total", ]$Df + 1
    MSe = x["Residual", ]$SumOfSqs/x["Residual", ]$Df
    out = x %>% as.data.frame %>% rownames_to_column(var = "Variable") %>% mutate(Omega2_partial = ifelse(is.na(`F`), NA, (SumOfSqs - Df * MSe)/(SumOfSqs + (N_t - Df) * MSe)))
    return(out)
}

TimePoint_color <- c("Week3"="#c7e9b4", "Week6"="#7fcdbb", 
                     "Week9-14"="#41b6c4", "Week18"="#2c7fb8")
```

## over all explore
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta = meta_all %>% filter(SampleID %in% labels(wdist))
  ## ape::pcoa
  wpcoa <- pcoa(wdist)
  varPC1 <- round(wpcoa$values$Relative_eig[1]*100, 2)
  varPC2 <- round(wpcoa$values$Relative_eig[2]*100, 2)
  wdat <- merge(wpcoa$vectors[,1:2], wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = `Axis.1`, PC2 = `Axis.2`) %>% 
    mutate(grp1 = ifelse(is.na(BirthMode),"Mom-vagina", str_c("Pup", BirthMode, BodySite, sep = "-")),
           grp2 = ifelse(is.na(BirthMode),"Mom-vagina-day0", str_c("Pup", BodySite,TimePoint, sep = "-"))) %>% 
    mutate(grp1 = factor(grp1, levels = c("Pup-CS-Feces","Pup-CSR-Feces","Pup-VF-Feces","Mom-vagina"), ordered = T),
           grp2 = factor(grp2, levels = c("Mom-vagina-day0","Pup-Feces-Week3","Pup-Feces-Week6","Pup-Feces-Week9-14","Pup-Feces-Week18"), ordered = T))
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  plot_g1 <- function(wdat, x_lab, y_lab){
    ggplot(wdat, aes(x = PC1, y = PC2, color = grp1)) + 
      geom_point() + 
      scale_color_manual(values = c("Pup-CS-Feces"="#E41A1C", "Pup-CSR-Feces"="#4DAF4A",
                                    "Pup-VF-Feces"="#377EB8", "Mom-vagina"="Gray50"),name = "") + 
      labs(x = x_lab, y = y_lab, title = "") + 
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold')) 
  }
  
  g <- plot_g1(wdat, x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_birthmode_", dd, "_ape.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  plot_g2 <- function(wdat, x_lab, y_lab){
    ggplot(wdat, aes(x = PC1, y = PC2, color = grp2)) + 
      geom_point() + 
      scale_color_manual(values = c("Mom-vagina-day0"="Gray50", "Pup-Feces-Week3"="#c7e9b4", 
                                    "Pup-Feces-Week6"="#7fcdbb", "Pup-Feces-Week9-14"="#41b6c4",
                                    "Pup-Feces-Week18"="#2c7fb8"),name = "") + 
      labs(x = x_lab, y = y_lab, title = "") + 
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold')) 
  }
  
  g <- plot_g2(wdat, x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_TimePoint_", dd, "_ape.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  
  ## vegan::dbrda
  wrda <- dbrda(wdist ~ 1)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig[1]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig[2]/sum(wrda_eig)*100,2)
  #wrda_fit <- envfit(wrda ~ BodySite + Sex, data = wmeta)
  #ordiplot(wrda, display = "site")
  #plot(wrda_fit)
  wdat <- merge(wrda$CA$u[,1:2], wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2) %>% 
    mutate(grp1 = ifelse(is.na(BirthMode),"Mom-vagina", str_c("Pup", BirthMode, BodySite, sep = "-")),
           grp2 = ifelse(is.na(BirthMode),"Mom-vagina-day0", str_c("Pup", BodySite,TimePoint, sep = "-"))) %>% 
    mutate(grp1 = factor(grp1, levels = c("Pup-CS-Feces","Pup-CSR-Feces","Pup-VF-Feces","Mom-vagina"), ordered = T),
           grp2 = factor(grp2, levels = c("Mom-vagina-day0","Pup-Feces-Week3","Pup-Feces-Week6","Pup-Feces-Week9-14","Pup-Feces-Week18"), ordered = T))
  
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  g <- plot_g1(wdat,  x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_birthmode_", dd, "_vegan.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  g <- plot_g2(wdat,  x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_TimePoint_", dd, "_vegan.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  
  ## vegan::metaMDS
  set.seed(1)
  wnmds <- metaMDS(wdist, k = 2, try = 50, trymax = 100)
  wnmds_fit <- envfit(wnmds ~ BodySite + Sex, data = wmeta)
  #ordiplot(wnmds, display = "site")
  #plot(wnmds_fit)
  wdat <- merge(scores(wnmds,display = "site"), wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = NMDS1, PC2 = NMDS2) %>% 
    mutate(grp1 = ifelse(is.na(BirthMode),"Mom-vagina", str_c("Pup", BirthMode, BodySite, sep = "-")),
           grp2 = ifelse(is.na(BirthMode),"Mom-vagina-day0", str_c("Pup", BodySite,TimePoint, sep = "-"))) %>% 
    mutate(grp1 = factor(grp1, levels = c("Pup-CS-Feces","Pup-CSR-Feces","Pup-VF-Feces","Mom-vagina"), ordered = T),
           grp2 = factor(grp2, levels = c("Mom-vagina-day0","Pup-Feces-Week3","Pup-Feces-Week6","Pup-Feces-Week9-14","Pup-Feces-Week18"), ordered = T))
  x_lab <- "NMDS1"
  y_lab <- "NMDS2"
  g <- plot_g1(wdat,  x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_birthmode_", dd, "_NMDS.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  g <- plot_g2(wdat,  x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/vaginal_fecal_by_TimePoint_", dd, "_NMDS.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
}
```

## effect of birthmode, by sex and TimePoint
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta <- meta_all %>% filter(SampleID %in% labels(wdist)) %>% 
    filter(Generation == "F1")
  wdist <- dist_subset(wdist, wmeta$SampleID)
  
  all_comb_df <- expand.grid(TimePoint = c("All", "Week3", "Week6", "Week9-14", "Week18"),
                             Sex = c("All", "M", "F"),
                             stringsAsFactors = F)
  comp_BirthMode <- combn(wmeta$BirthMode %>% unique() %>% as.character(),2)
  comp_TimePoint <- combn(wmeta$TimePoint %>% unique() %>% as.character(),2)
  
  res_dispersion <- list()
  res_permanova <- list()
  
  for(i in 1:nrow(all_comb_df)){
    if(all_comb_df[i,"TimePoint"] ==  "All" & all_comb_df[i,"Sex"] ==  "All"){
      # dispersion all variable
      disper <- betadisper(wdist, wmeta$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist, wmeta$Sex)
      d2 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d2 <- d2 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "Sex")
      disper <- betadisper(wdist, wmeta$TimePoint)
      d3 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d3 <- d3 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "TimePoint")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d2, d3))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(i){
        wmeta_ss <- wmeta %>% filter(BirthMode %in% comp_BirthMode[,i])
        wdist_ss <- dist_subset(wdist, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex + TimePoint, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,i], comp_BirthMode[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint") 
      # permanova pairwise TimePoint
      res2 <- lapply(1:ncol(comp_TimePoint), function(i){
        wmeta_ss <- wmeta %>% filter(TimePoint %in% comp_TimePoint[,i])
        wdist_ss <- dist_subset(wdist, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex + TimePoint, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_TimePoint[1,i], comp_TimePoint[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint") 
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist ~ BirthMode + Sex + TimePoint, data = wmeta, by = "margin")) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint")
      res_permanova <- list.append(res_permanova, rbind(res1, res2, res3))
    }else if(all_comb_df[i,"TimePoint"] !=  "All" & all_comb_df[i,"Sex"] ==  "All"){
      wmeta_s <- wmeta %>% filter(TimePoint == all_comb_df[i,"TimePoint"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist, wmeta$Sex)
      d2 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d2 <- d2 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "Sex")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d2))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(i){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,i])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,i], comp_BirthMode[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex") 
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode + Sex, data = wmeta_s, by = "margin")) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex") 
      res_permanova <- list.append(res_permanova, rbind(res1, res3))     
    }else if(all_comb_df[i,"TimePoint"] ==  "All" & all_comb_df[i,"Sex"] !=  "All"){
      wmeta_s <- wmeta %>% filter(Sex == all_comb_df[i,"Sex"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist_s, wmeta_s$TimePoint)
      d3 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d3 <- d3 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "TimePoint")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d3))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(i){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,i])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + TimePoint, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,i], comp_BirthMode[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint") 
      # permanova pairwise TimePoint      
      res2 <- lapply(1:ncol(comp_TimePoint), function(i){
        wmeta_ss <- wmeta_s %>% filter(TimePoint %in% comp_TimePoint[,i])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode +TimePoint, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_TimePoint[1,i], comp_TimePoint[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint") 
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode + TimePoint, data = wmeta_s, by = "margin")) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint") 
      res_permanova <- list.append(res_permanova, rbind(res1, res2, res3))    
    }else if(all_comb_df[i,"TimePoint"] !=  "All" & all_comb_df[i,"Sex"] !=  "All"){
      wmeta_s <- wmeta %>% filter(Sex == all_comb_df[i,"Sex"], TimePoint == all_comb_df[i,"TimePoint"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      res_dispersion <- list.append(res_dispersion, d1)
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(j){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,j])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode, data = wmeta_ss, by = "margin")) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,j], comp_BirthMode[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode")
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode, data = wmeta_s, by = "margin")) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode") 
      res_permanova <- list.append(res_permanova, rbind(res1, res3))    
    }  
  }
  res <- res_permanova %>% do.call(rbind, .) 
  res$padj <- p.adjust(res$`Pr(>F)`, method = "fdr")
  write.csv(res,
            file = paste0(temp_out, "beta/adonis_all_",dd, "_table.csv"),
            quote = F, row.names = F)
  write.csv(res_dispersion %>% do.call(rbind, .),
            file = paste0(temp_out, "beta/dispersion_all_",dd, "_table.csv"),
            quote = F, row.names = F)
  
}
```

# effect of birthmode, by sex and TimePoint adjusted by FamilyID 
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta <- meta_all %>% filter(SampleID %in% labels(wdist)) %>% 
    filter(Generation == "F1")
  wdist <- dist_subset(wdist, wmeta$SampleID)
  
  all_comb_df <- expand.grid(TimePoint = c("All", "Week3", "Week6", "Week9-14", "Week18"),
                             Sex = c("All", "M", "F"),
                             stringsAsFactors = F)
  comp_BirthMode <- combn(wmeta$BirthMode %>% unique() %>% as.character(),2)
  comp_TimePoint <- combn(wmeta$TimePoint %>% unique() %>% as.character(),2)
  
  res_dispersion <- list()
  res_permanova <- list()
  
  for(i in 1:nrow(all_comb_df)){
    if(all_comb_df[i,"TimePoint"] ==  "All" & all_comb_df[i,"Sex"] ==  "All"){
      # dispersion all variable
      disper <- betadisper(wdist, wmeta$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist, wmeta$Sex)
      d2 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d2 <- d2 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "Sex")
      disper <- betadisper(wdist, wmeta$TimePoint)
      d3 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d3 <- d3 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "TimePoint")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d2, d3))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(j){
        wmeta_ss <- wmeta %>% filter(BirthMode %in% comp_BirthMode[,j])
        wdist_ss <- dist_subset(wdist, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex + TimePoint, data = wmeta_ss, by = "margin", strata = wmeta_ss$FamilyID)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,j], comp_BirthMode[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint|FamilyID") 
      # permanova pairwise TimePoint
      res2 <- lapply(1:ncol(comp_TimePoint), function(j){
        wmeta_ss <- wmeta %>% filter(TimePoint %in% comp_TimePoint[,j])
        wdist_ss <- dist_subset(wdist, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex + TimePoint, data = wmeta_ss, by = "margin", strata = wmeta_ss$FamilyID)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_TimePoint[1,j], comp_TimePoint[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint|FamilyID") 
      # permanova
      res3 <- calcOmega2_dm(adonis2(wdist ~ BirthMode + Sex + TimePoint, data = wmeta, by = "margin", strata = wmeta$FamilyID)) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex + TimePoint|FamilyID")
      res_permanova <- list.append(res_permanova, rbind(res1, res2, res3))
    }else if(all_comb_df[i,"TimePoint"] !=  "All" & all_comb_df[i,"Sex"] ==  "All"){
      wmeta_s <- wmeta %>% filter(TimePoint == all_comb_df[i,"TimePoint"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist, wmeta$Sex)
      d2 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d2 <- d2 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "Sex")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d2))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(j){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,j])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + Sex, data = wmeta_ss, by = "margin",strata = wmeta_ss$FamilyID)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,j], comp_BirthMode[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex|FamilyID") 
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode + Sex, data = wmeta_s, by = "margin", strata = wmeta_s$FamilyID)) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + Sex|FamilyID") 
      res_permanova <- list.append(res_permanova, rbind(res1, res3))     
    }else if(all_comb_df[i,"TimePoint"] ==  "All" & all_comb_df[i,"Sex"] !=  "All"){
      wmeta_s <- wmeta %>% filter(Sex == all_comb_df[i,"Sex"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      disper <- betadisper(wdist_s, wmeta_s$TimePoint)
      d3 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d3 <- d3 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "TimePoint")
      res_dispersion <- list.append(res_dispersion, rbind(d1, d3))
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(j){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,j])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode + TimePoint, data = wmeta_ss, by = "margin", strata = wmeta_ss$FamilyID)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,j], comp_BirthMode[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint|FamilyID") 
      # permanova pairwise TimePoint      
      res2 <- lapply(1:ncol(comp_TimePoint), function(j){
        wmeta_ss <- wmeta_s %>% filter(TimePoint %in% comp_TimePoint[,j])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode +TimePoint, data = wmeta_ss, by = "margin", strata = wmeta_ss$FamilyID)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_TimePoint[1,j], comp_TimePoint[2,j], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint|FamilyID") 
      # permanova 
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode + TimePoint, data = wmeta_s, by = "margin",strata = wmeta_s$FamilyID)) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode + TimePoint") 
      res_permanova <- list.append(res_permanova, rbind(res1, res2, res3))    
    }else if(all_comb_df[i,"TimePoint"] !=  "All" & all_comb_df[i,"Sex"] !=  "All"){
      wmeta_s <- wmeta %>% filter(Sex == all_comb_df[i,"Sex"], TimePoint == all_comb_df[i,"TimePoint"])
      wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
      # dispersion BirthMode, Sex
      disper <- betadisper(wdist_s, wmeta_s$BirthMode)   
      d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
        rownames_to_column(var = "Comp")
      d1 <- d1 %>% mutate(Sex =  all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], Variable = "BirthMode")
      res_dispersion <- list.append(res_dispersion, d1)
      # permanova pairwise BirthMode
      res1 <- lapply(1:ncol(comp_BirthMode), function(i){
        wmeta_ss <- wmeta_s %>% filter(BirthMode %in% comp_BirthMode[,i])
        wdist_ss <- dist_subset(wdist_s, wmeta_ss$SampleID)
        perms <- with(wmeta_ss, how(blocks = FamilyID, complete = FALSE, observed = TRUE))
        calcOmega2_dm(adonis2(wdist_ss ~ BirthMode, data = wmeta_ss, by = "margin", permutations = perms)) %>% 
          as.data.frame() %>% mutate(Comp = str_c(comp_BirthMode[1,i], comp_BirthMode[2,i], sep = "-")) %>% 
          filter(!is.na(F))
      }) %>% do.call(rbind, .) %>% 
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode|FamilyID")
      # permanova 
      perms <- with(wmeta_s, how(blocks = FamilyID, complete = FALSE, observed = TRUE))
      res3 <- calcOmega2_dm(adonis2(wdist_s ~ BirthMode, data = wmeta_s, by = "margin", permutations = perms)) %>% 
        as.data.frame() %>% mutate(Comp = "All") %>% 
        filter(!is.na(F)) %>%  
        mutate(Sex = all_comb_df[i,"Sex"], TimePoint = all_comb_df[i,"TimePoint"], fo = "~ BirthMode|FamilyID") 
      res_permanova <- list.append(res_permanova, rbind(res1, res3))    
    }  
  }
  res <- res_permanova %>% do.call(rbind, .) 
  res$padj <- p.adjust(res$`Pr(>F)`, method = "fdr")
  write.csv(res,
            file = paste0(temp_out, "beta/adonis_all_",dd, "_adj_FamilyID_table.csv"),
            quote = F, row.names = F)
  write.csv(res_dispersion %>% do.call(rbind, .),
            file = paste0(temp_out, "beta/dispersion_all_",dd, "_adj_FamilyID_table.csv"),
            quote = F, row.names = F)
}


 
```

## plot pups PCoA use vegan mds, nmds
```{r}
for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta <- meta_all %>% filter(SampleID %in% labels(wdist)) %>% 
    filter(Generation == "F1")
  wdist <- dist_subset(wdist, wmeta$SampleID)
  
  wrda <- dbrda(wdist ~ 1)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig[1]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig[2]/sum(wrda_eig)*100,2)
  wrda_fit <- envfit(wrda ~ BirthMode + Sex + TimePoint, data = wmeta)
  wdat_rda <- merge(wrda$CA$u[,1:2], wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  #Pcoa TimePoint
  plot_Age_group_pcoa <- function(wdat, x_lab, y_lab){
    wdat_c <- wdat %>% group_by(TimePoint, Sex) %>% summarise(across(starts_with("PC"), mean)) %>% 
      ungroup()
    wdat_2 <- merge(wdat, wdat_c, by = c("Sex", "TimePoint"), suffixes = c("",".c"))
    ggplot(wdat_2, aes(x = PC1, y = PC2, color = TimePoint)) + 
      geom_segment(aes(x = PC1, xend  = PC1.c, y = PC2, yend = PC2.c), size = 0.3) + 
      #geom_point(aes(shape = BirthMode), size = 2) + 
      geom_point(size = 1, alpha = 0.7) + 
      geom_point(data = wdat_c, size = 5, alpha = 0.7, aes(fill = TimePoint)) + 
      #stat_ellipse(type = "t") + 
      scale_color_manual(values = TimePoint_color) + 
      labs(x = x_lab, y = y_lab, title = "") + 
      facet_wrap(~ Sex, ncol = 2, scales = "fixed", labeller = labeller(Sex = Sex_labs)) +
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold'))    
  }
  g <- plot_Age_group_pcoa(wdat_rda, x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_pcoa.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  #PC1 by TimePoint Birthmode
  plot_Age_group_pc1 <- function(wdat, x_lab){
    ggplot(wdat, aes(x = TimePoint, y = PC1)) + 
      geom_boxplot(aes(color=BirthMode, fill=BirthMode), alpha = 0.5, outlier.alpha = 0) + 
      geom_point(aes(color=BirthMode), position = position_dodge2(width = 0.6)) + 
      stat_compare_means(aes(group = BirthMode), hide.ns = T, label =  "p.signif") +
      scale_color_manual(values = BirthMode_color, aesthetics = c("color","fill")) + 
      labs(x="",y=x_lab) + 
      facet_wrap(~ Sex, ncol = 2, scales = "fixed", labeller = labeller(Sex = Sex_labs)) +
      coord_flip() +
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold'))
  }
  g <- plot_Age_group_pc1(wdat_rda, x_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_pc1.pdf"), device = cairo_pdf, 
         plot = g, width = 10, height = 5, units = "in", dpi = 300)
  #PC2 by TimePoint Birthmode
  plot_Age_group_pc2 <- function(wdat, y_lab){
    g <- ggplot(wdat, aes(x = TimePoint, y = PC2)) + 
      geom_boxplot(aes(color=BirthMode, fill=BirthMode), alpha = 0.5, outlier.alpha = 0) + 
      geom_point(aes(color=BirthMode), position = position_dodge2(width = 0.6)) + 
      stat_compare_means(aes(group = BirthMode), hide.ns = T, label =  "p.signif") +
      scale_color_manual(values = BirthMode_color, aesthetics = c("color","fill")) + 
      labs(x="",y=y_lab) + 
      facet_wrap(~ Sex, ncol = 2, scales = "fixed", labeller = labeller(Sex = Sex_labs)) +
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold'))
  }
  g <- plot_Age_group_pc2(wdat_rda, y_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_pc2.pdf"), device = cairo_pdf, 
         plot = g, width = 10, height = 5, units = "in", dpi = 300)  
  ##Poca birthmode by TimePoint
  plot_birthmode_pcoa_Age_group <- function(wdat, x_lab, y_lab){
    wdat_c <- wdat %>% group_by(BirthMode, TimePoint, Sex) %>% summarise(across(starts_with("PC"), mean)) %>% 
      ungroup()
    wdat_2 <- merge(wdat, wdat_c, by = c("Sex", "TimePoint", "BirthMode"), suffixes = c("",".c"))
    ggplot(wdat_2, aes(x = PC1, y = PC2, color = BirthMode)) +     
      geom_segment(aes(x = PC1, xend = PC1.c, y = PC2, yend = PC2.c), size = 0.3) + 
      geom_point(size = 1, alpha = 0.7) + 
      geom_point(data = wdat_c, size = 5, alpha = 0.7, aes(fill = BirthMode)) + 
      #stat_ellipse(type = "t") + 
      scale_color_manual(values = BirthMode_color) + 
      labs(x = x_lab, y = y_lab, title = "") + 
      facet_grid(TimePoint~Sex,
                 labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold'))
  }
  g <- plot_birthmode_pcoa_Age_group(wdat_rda, x_lab, y_lab)   
  ggsave(filename = paste0(temp_out,"beta/pup_by_birthmode_TimePoint_",dd,"_pcoa.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 10, units = "in", dpi = 300)
  ##Poca sex by TimePoint
  plot_sex_pcoa_Age_group <- function(wdat, x_lab, y_lab){
    wdat_c <- wdat %>% group_by(TimePoint, Sex) %>% summarise(across(starts_with("PC"), mean)) %>% 
      ungroup()
    wdat_2 <- merge(wdat, wdat_c, by = c("Sex", "TimePoint"), suffixes = c("",".c"))
    ggplot(wdat_2, aes(x = PC1, y = PC2, color = Sex)) +     
      geom_segment(aes(x = PC1, xend = PC1.c, y = PC2, yend = PC2.c), size = 0.3) + 
      geom_point(size = 1, alpha = 0.7) + 
      geom_point(data = wdat_c, size = 5, alpha = 0.7, aes(fill = Sex)) + 
      scale_color_manual(aesthetics = c("color","fill"),values = Sex_color, labels = Sex_labs) + 
      labs(x = x_lab, y = y_lab, title = "") + 
      facet_wrap(~TimePoint, nrow = 2) + 
      theme_bw() + theme(aspect.ratio = 1,
                         panel.background = element_rect(fill = NA),
                         strip.text = element_text(size=12,color="black",face='bold'),
                         axis.title = element_text(size=12,color="black",face='bold'),
                         axis.text = element_text(size=12,color="black",face='bold'),
                         axis.text.x = element_text(size=12,color="black",face='bold'),
                         axis.text.y = element_text(size=12,color="black",face='bold'),
                         legend.text = element_text(size=12,color="black",face='bold'),
                         legend.title = element_text(size=12,color="black",face='bold'),
                         title = element_text(size=12,color="black",face='bold'))
  }
  g <- plot_sex_pcoa_Age_group(wdat_rda, x_lab, y_lab)   
  ggsave(filename = paste0(temp_out,"beta/pup_by_sex_TimePoint_",dd,"_pcoa.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  #explore birthmode effect on PCs
  wdat <- merge(wrda$CA$u[,1:9], wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    dplyr::select(BirthMode, Sex, TimePoint, MDS1:MDS9) %>% 
    pivot_longer(MDS1:MDS9, names_to = "PC")
  
  stat.test <- wdat %>% group_by(Sex, TimePoint, PC) %>% 
    rstatix::wilcox_test(value ~ BirthMode) %>% 
    rstatix::adjust_pvalue(method = "fdr") %>%
    rstatix::add_significance("p.adj") %>% 
    rstatix::add_xy_position(x = "TimePoint", dodge = 0.8, scales = "free_y") %>% 
    filter(!p.adj.signif=="ns") 
  g <- ggplot(wdat, aes(x = TimePoint, y = value, )) + 
    geom_boxplot(aes(fill=BirthMode)) + 
    #stat_compare_means(aes(group=BirthMode), hide.ns = T, label = "p.format") + 
    stat_pvalue_manual(stat.test, label = "p.adj.signif", tip.length = 0, size = 12/.pt, hide.ns = T) +
    scale_color_manual(values = BirthMode_color) +
    labs(x="", y="", title = "") + 
    facet_grid(Sex~PC, labeller = labeller(Sex = Sex_labs)) + 
    theme(axis.text.x = element_text(angle = 90))
  ggsave(filename = paste0(temp_out,"beta/pup_by_birthmode_TimePoint_",dd,"_9D.pdf"), device = cairo_pdf, 
         plot = g, width = 14, height = 7, units = "in", dpi = 300) 
  #dbRDA
  wrda2 <- dbrda(wdist ~ BirthMode + Sex + TimePoint, data = wmeta)
  wrda_eig2 <- eigenvals(wrda2)
  varRDA1 <- round(wrda_eig2["dbRDA1"]/sum(wrda_eig2)*100,2)
  varRDA2 <- round(wrda_eig2["dbRDA2"]/sum(wrda_eig2)*100,2)
  x_lab <- paste0("dbRDA1 (", varRDA1, "%)")
  y_lab <- paste0("dbRDA2 (", varRDA2, "%)")
  wrda_score2 <- scores(wrda2)    
  wdat_rda2 <- merge(wrda_score2$sites,
                      wmeta, by.x = 0, by.y = "SampleID", all.x = T)
  wdat_center_BirthMode <- wrda_score2$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "rowname") %>%
    filter(str_detect(rowname, "^Birth")) %>% 
    transmute(BirthMode = str_remove(rowname, "BirthMode"), 
              dbRDA1.birth = dbRDA1, dbRDA2.birth = dbRDA2) 
  wdat_center_Sex <- wrda_score2$centroids %>% as.data.frame() %>% 
    rownames_to_column(var = "rowname") %>%
    filter(str_detect(rowname, "^Sex")) %>% 
    transmute(Sex = str_remove(rowname, "Sex"), 
              dbRDA1.sex = dbRDA1, dbRDA2.sex = dbRDA2) 
  wdat <- wdat_rda2 %>% 
    left_join(wdat_center_Sex, by = "Sex") %>%
    left_join(wdat_center_BirthMode, by = "BirthMode") %>% 
    rename(PC1 = dbRDA1, PC2 = dbRDA2)
  g <- plot_Age_group_pcoa(wdat, x_lab, y_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_dbRDA.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  g <- plot_birthmode_pcoa_Age_group(wdat, x_lab, y_lab)   
  ggsave(filename = paste0(temp_out,"beta/pup_by_birthmode_TimePoint_",dd,"_dbRDA.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 10, units = "in", dpi = 300)
  g <- plot_sex_pcoa_Age_group(wdat, x_lab, y_lab)   
  ggsave(filename = paste0(temp_out,"beta/pup_by_sex_TimePoint_",dd,"_dbRDA.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  # NMDS
  set.seed(1)
  wnmds <- metaMDS(wdist, k = 2, try = 50, trymax = 100)
  wnmds_fit <- envfit(wnmds ~ BirthMode + Sex + TimePoint, data = wmeta)
  wdat_nmds <- merge(scores(wnmds,display = "site"), wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = NMDS1, PC2 = NMDS2) 
  x_lab <- "NMDS1"
  y_lab <- "NMDS2"
  g <- plot_Age_group_pcoa(wdat_nmds, x_lab, y_lab)
  ggsave(filename = paste0(temp_out, "beta/pup_by_TimePoint_", dd, "_nmds.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)  
  g <- plot_Age_group_pc1(wdat_nmds, x_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_nmds1.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  g <- plot_Age_group_pc2(wdat_nmds, y_lab)
  ggsave(filename = paste0(temp_out,"beta/pup_by_TimePoint_", dd, "_nmds2.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)      
  g <- plot_birthmode_pcoa_Age_group(wdat_nmds, x_lab, y_lab)   
  ggsave(filename = paste0(temp_out,"beta/pup_by_birthmode_TimePoint_", dd,"_nmds.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
  g <- plot_sex_pcoa_Age_group(wdat_nmds, x_lab, y_lab) 
  ggsave(filename = paste0(temp_out,"beta/pup_by_sex_TimePoint_", dd, "_nmds.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 5, units = "in", dpi = 300)
}
```


## overweight
```{r}
wmeta <- meta_all %>% left_join(at_week15 %>% dplyr::select(MouseID, BW_status, Overweight, Underweight), 
                                by = "MouseID") %>% filter(!is.na(BW_status))

disper_l <- list()
adoins_l <- list()

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta <- wmeta %>% filter(SampleID %in% labels(wdist)) 
  wdist <- dist_subset(wdist, wmeta$SampleID)
  
  disper <- betadisper(wdist, wmeta$Overweight)
  disper_l[dd] <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
    rownames_to_column(var = "Variable")
  
  per <- adonis2(wdist ~ Overweight+BirthMode+TimePoint+Sex, data = wmeta, by = "margin")
  adoins_l[dd] = calcOmega2_dm(per)
  
  ## by TimePoint and Sex
  lst_df <- expand.grid(wmeta$Sex %>% unique(), 
                        wmeta$TimePoint %>% unique(), 
                        stringsAsFactors = F)
  names(lst_df) <- c("Sex", "TimePoint")
  res_sum <- lapply(1:nrow(lst_df), function(i){
    wmeta_s <- wmeta %>% filter(Sex == lst_df[i,"Sex"], TimePoint == lst_df[i,"TimePoint"])
    wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
    disper <- betadisper(wdist_s, wmeta_s$Overweight)   
    d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
      rownames_to_column(var = "Comp")
    d1 <- d1 %>% mutate(Sex = lst_df[i,"Sex"], TimePoint = lst_df[i,"TimePoint"], Variable = "Overweight")
    
    res <- calcOmega2_dm(adonis2(wdist_s ~ Overweight, data = wmeta_s, by = "terms")) %>% 
      as.data.frame() %>% mutate(Comp = "Overweight") %>% 
      mutate(Sex = lst_df[i,"Sex"], TimePoint = lst_df[i,"TimePoint"], fo = "~ Overweight")    
    list(dispersion = d1, permanova = res)
  })
  res_dispersion_1 <- lapply(1:length(res_sum),function(i){res_sum[[i]]$dispersion}) %>% do.call(rbind,.)
  res_permanova_1 <- lapply(1:length(res_sum),function(i){res_sum[[i]]$permanova}) %>% do.call(rbind,.)
  ## by TimePoint
  res_sum <- lapply(wmeta$TimePoint %>% unique(), function(tp){
    wmeta_s <- wmeta %>% filter(TimePoint == tp)
    wdist_s <- dist_subset(wdist, wmeta_s$SampleID)
    disper <- betadisper(wdist_s, wmeta_s$Overweight)   
    d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
      rownames_to_column(var = "Comp")
    d1 <- d1 %>% mutate(Sex = "Both", TimePoint = tp, Variable = "Overweight")
    
    res <- calcOmega2_dm(adonis2(wdist_s ~ Overweight+Sex, data = wmeta_s, by = "terms")) %>% 
      as.data.frame() %>% mutate(Comp = "Overweight") %>% 
      mutate(Sex = "Both", TimePoint = tp, fo = "~ Overweight+Sex")    
    list(dispersion = d1, permanova = res)
  })
  res_dispersion_2 <- lapply(1:length(res_sum),function(i){res_sum[[i]]$dispersion}) %>% do.call(rbind,.)
  res_permanova_2 <- lapply(1:length(res_sum),function(i){res_sum[[i]]$permanova}) %>% do.call(rbind,.)
  ## 
  disper <- betadisper(wdist, wmeta$Overweight)   
  d1 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
    rownames_to_column(var = "Comp")
  d1 <- d1 %>% mutate(Sex = "Both", TimePoint = "All", Variable = "Overweight")
  disper <- betadisper(wdist, wmeta$Sex)
  d2 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
    rownames_to_column(var = "Comp")
  d2 <- d2 %>% mutate(Sex = "Both", TimePoint = "All", Variable = "Sex")
  disper <- betadisper(wdist, wmeta$TimePoint)
  d3 <- (TukeyHSD(disper, which = "group"))$group %>% as.data.frame() %>% 
    rownames_to_column(var = "Comp")
  d3 <- d3 %>% mutate(Sex = "Both", TimePoint = "All", Variable = "TimePoint")
  res_dispersion_3 <- rbind(d1, d2, d3)
  
  res_permanova_3 <- calcOmega2_dm(adonis2(wdist ~ Overweight+BirthMode+TimePoint+Sex, data = wmeta, by = "terms")) %>% 
    as.data.frame() %>% mutate(Comp = "All") %>% filter(!is.na(F)) %>% 
    mutate(Sex = "Both", TimePoint = "All", fo = "~ BirthMode + Sex + TimePoint") 
  
  res <- rbind(res_permanova_1, res_permanova_2, res_permanova_3)
  res$padj <- p.adjust(res$`Pr(>F)`, method = "fdr") 
  write.csv(res %>% filter(!is.na(`F`)),
            file = paste0(temp_out, "beta/adonis_Overweight_",dd, "_table.csv"),
            quote = F, row.names = F)
  write.csv(rbind(res_dispersion_1, res_dispersion_2, res_dispersion_3),
            file = paste0(temp_out, "beta/dispersion_Overweight_",dd, "_table.csv"),
            quote = F, row.names = F)
}

for(dd in c("uu","wu","bray","jaccard")){
  wdist <- get(paste0("DM_",dd))
  wmeta <- wmeta %>% filter(SampleID %in% labels(wdist)) 
  wdist <- dist_subset(wdist, wmeta$SampleID)
  ## pcoa
  wrda <- dbrda(wdist ~ 1)
  wrda_eig <- eigenvals(wrda)
  varPC1 <- round(wrda_eig[1]/sum(wrda_eig)*100,2)
  varPC2 <- round(wrda_eig[2]/sum(wrda_eig)*100,2)
  wrda_fit <- envfit(wrda ~ Overweight + BirthMode + Sex + TimePoint, data = wmeta)
  wdat_rda <- merge(wrda$CA$u[,1:2], wmeta, by.x = 0, by.y = "SampleID", all.x = T) %>% 
    rename(PC1 = MDS1, PC2 = MDS2)
  x_lab <- paste0("PC1 (", varPC1, "%)")
  y_lab <- paste0("PC2 (", varPC2, "%)")
  
  wdat_c <- wdat_rda %>% group_by(TimePoint, Sex, Overweight) %>% summarise(across(starts_with("PC"), mean)) %>% 
    ungroup()
  wdat_2 <- merge(wdat_rda, wdat_c, by = c("TimePoint", "Sex", "Overweight"), suffixes = c("",".c"))
  
  g <- ggplot(wdat_2, aes(x = PC1, y = PC2, color = Overweight)) + 
    geom_point(size = 1, alpha = 0.7) + 
    geom_segment(aes(x = PC1, xend = PC1.c, y = PC2, yend = PC2.c), size = 0.3) + 
    geom_point(data = wdat_c, size = 5, alpha = 0.7, aes(fill = Overweight)) + 
    stat_ellipse(type = "t") + 
    scale_color_manual(values = c("Y"="#d73027", "N"="#1a9850"), 
                       name = "Overweight") + 
    labs(x = x_lab, y = y_lab, title = "") + 
    facet_grid(TimePoint~Sex,
               labeller = labeller(Sex = Sex_labs)) + 
    theme_bw() + theme(aspect.ratio = 1,
                       panel.background = element_rect(fill = NA),
                       strip.text = element_text(size=12,color="black",face='bold'),
                       axis.title = element_text(size=12,color="black",face='bold'),
                       axis.text = element_text(size=12,color="black",face='bold'),
                       axis.text.x = element_text(size=12,color="black",face='bold'),
                       axis.text.y = element_text(size=12,color="black",face='bold'),
                       legend.text = element_text(size=12,color="black",face='bold'),
                       legend.title = element_text(size=12,color="black",face='bold'),
                       title = element_text(size=12,color="black",face='bold'))
  ggsave(filename = paste0(temp_out,"beta/pup_Overweight_by_TimePoint_", dd, "_pcoa.pdf"), device = cairo_pdf, 
         plot = g, width = 7, height = 10, units = "in", dpi = 300)
}


```
