==> code/05_source_track.Rmd <==
---
title: "Source track"
author: "Haipeng Sun"
date: "7/14/2022"
output: html_document
---

## library
```{r setup, message=FALSE, warning=FALSE}
library(knitr)
library(phyloseq)
library(usedist)
library(purrr)
library(furrr)
library(vegan)
library(ape)
library(rlist)
library(ggpubr)
library(tidyverse)
```


## load data
```{r}
load(file = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/data/BW_and_phylo.RData")
#load(file = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/data/BW_responder.RData")
temp_out = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/output/sourcetrack/"
```

## Source track FEAST
### run FEAST
```{r}
library(FEAST)
wd <- getwd()
otu <- otu_table(phylo_decontam_clean)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>%
  mutate(Env = ifelse(Generation == "F0", "Mom_Vaginal", "Pup_fecal"),
         SourceSink = ifelse(Generation == "F0", "Source", "Sink"),
         id = FamilyNo) %>% 
  dplyr::select(SampleID, Env, SourceSink, id) %>% 
  arrange(id, desc(Env))

for (fid in Family_id){
  wdat <- dat %>% filter(id == fid)
  wdat$id <- c(1:nrow(wdat)) 
  wdat <- wdat %>% mutate(id = ifelse(SourceSink == "Source", NA, id)) %>%
      column_to_rownames(var = "SampleID")
  ft <- otu %>% dplyr::select(id, row.names(wdat)) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  FEAST_output <- FEAST(C = ft, metadata = wdat, different_sources_flag = 0, 
                       dir_path = temp_out, 
                       outfile=paste0(fid,"_no_rf"))
  setwd(wd)
}

otu <- otu_table(phylo_decontam_clean_rf)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>%
  mutate(Env = ifelse(Generation == "F0", "Mom_Vaginal", "Pup_fecal"),
         SourceSink = ifelse(Generation == "F0", "Source", "Sink"),
         id = FamilyNo) %>% 
  dplyr::select(SampleID, Env, SourceSink, id) %>% 
  arrange(id, desc(Env))

for (fid in Family_id){
  wdat <- dat %>% filter(id == fid)
  wdat$id <- c(1:nrow(wdat)) 
  wdat <- wdat %>% mutate(id = ifelse(SourceSink == "Source", NA, id)) %>%
      column_to_rownames(var = "SampleID")
  ft <- otu %>% dplyr::select(id, row.names(wdat)) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  FEAST_output <- FEAST(C = ft, metadata = wdat, different_sources_flag = 0, 
                       dir_path = temp_out, 
                       outfile=paste0(fid,"_rf"))
  setwd(wd)
}
```

### load FEAST result
```{r}
otu <- otu_table(phylo_decontam_clean)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

res <- lapply(Family_id, function(fid){
  res <- read.table(file = paste0(temp_out,fid,"_no_rf_source_contributions_matrix.txt"))
  res <- res %>% rownames_to_column(var = "SampleID") %>% 
    mutate(SampleID = str_remove(SampleID, "_Pup_fecal"))
  names(res) <- c("SampleID", "Mom_vaginal", "Unknown")
  res
})
dat <- do.call(rbind, res) %>% left_join(meta_all, by = "SampleID")

g = ggplot(dat, aes(x = Age_group, y = Mom_vaginal)) + 
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom Vaginal Source Proportion (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"FEAST_by_family_vaginal_source_no_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)

otu <- otu_table(phylo_decontam_clean_rf)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

res <- lapply(Family_id, function(fid){
  res <- read.table(file = paste0(temp_out,fid,"_rf_source_contributions_matrix.txt"))
  res <- res %>% rownames_to_column(var = "SampleID") %>% 
    mutate(SampleID = str_remove(SampleID, "_Pup_fecal"))
  names(res) <- c("SampleID", "Mom_vaginal", "Unknown")
  res
})
dat <- do.call(rbind, res) %>% left_join(meta_all, by = "SampleID")

g = ggplot(dat, aes(x = Age_group, y = Mom_vaginal)) + 
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom Vaginal Source Proportion (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"FEAST_by_family_vaginal_source_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)
```

## Share ASV
```{r}
otu <- otu_table(phylo_decontam_clean)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()
dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>% arrange(FamilyNo, Generation)

res <- lapply(Family_id, function(fid){
  wdat <- dat %>% filter(FamilyNo == fid)
  ft <- otu %>% dplyr::select(id, wdat$SampleID) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  Mom_id <- wdat %>% filter(Generation == "F0") %>% pull(SampleID)
  ft_mom <- ft[pmatch(Mom_id,row.names(ft)),]
  Pup_id <- wdat %>% filter(Generation == "F1") %>% pull(SampleID)
  res <- lapply(Pup_id, function(pup){
    ft_pup <- ft[pmatch(pup,row.names(ft)),]
    mom_asv <- which(ft_mom>0) %>% names()
    pup_asv <- which(ft_pup>0) %>% names()
    shared_asv <- intersect(mom_asv, pup_asv)
    res <- data.frame(SampleID = pup,
                      Family_id = fid,
                      Mom_n_asv = length(mom_asv),
                      Pup_n_asv = length(pup_asv),
                      shared_n_asv = length(shared_asv),
                      shared_pup_prop = length(shared_asv) / length(pup_asv),
                      shared_pup_ra = sum(ft_pup[shared_asv]) / sum(ft_pup))
  })
  res <- do.call(rbind, res)  
})

dat <- do.call(rbind, res) %>% left_join(meta_all, by = "SampleID")
    
g = ggplot(dat, aes(x = Age_group, y = shared_pup_prop)) + 
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom ASV Proportion (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"ASV_share_prop_by_family_vaginal_source_no_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)

g = ggplot(dat, aes(x = Age_group, y = shared_pup_ra)) +     
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom ASV Relative abundance (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"ASV_share_ra_by_family_vaginal_source_no_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)

otu <- otu_table(phylo_decontam_clean_rf)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()
dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>% arrange(FamilyNo, Generation)

res <- lapply(Family_id, function(fid){
  wdat <- dat %>% filter(FamilyNo == fid)
  ft <- otu %>% dplyr::select(id, wdat$SampleID) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  Mom_id <- wdat %>% filter(Generation == "F0") %>% pull(SampleID)
  ft_mom <- ft[pmatch(Mom_id,row.names(ft)),]
  Pup_id <- wdat %>% filter(Generation == "F1") %>% pull(SampleID)
  res <- lapply(Pup_id, function(pup){
    ft_pup <- ft[pmatch(pup,row.names(ft)),]
    mom_asv <- which(ft_mom>0) %>% names()
    pup_asv <- which(ft_pup>0) %>% names()
    shared_asv <- intersect(mom_asv, pup_asv)
    res <- data.frame(SampleID = pup,
                      Family_id = fid,
                      Mom_n_asv = length(mom_asv),
                      Pup_n_asv = length(pup_asv),
                      shared_n_asv = length(shared_asv),
                      shared_pup_prop = length(shared_asv) / length(pup_asv),
                      shared_pup_ra = sum(ft_pup[shared_asv]) / sum(ft_pup))
  })
  res <- do.call(rbind, res)  
})

dat <- do.call(rbind, res) %>% left_join(meta_all, by = "SampleID")
    
g = ggplot(dat, aes(x = Age_group, y = shared_pup_prop)) + 
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom ASV Proportion (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"ASV_share_prop_by_family_vaginal_source_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)

g = ggplot(dat, aes(x = Age_group, y = shared_pup_ra)) +     
  geom_point(aes(color = FamilyID)) + 
  geom_line(aes(color = FamilyID,group = MouseID)) + 
  scale_y_continuous(breaks = c(0,0.2,0.4,0.6,0.8),labels = scales::percent) + 
  labs(x = "",y = "Mom ASV Relative abundance (%)", title = "") + 
  facet_grid(.~Sex, labeller = labeller(Sex = Sex_labs)) + 
      theme_bw() + theme(aspect.ratio = 0.8,
                           panel.background = element_rect(fill = NA),
                           strip.text = element_text(size=12,color="black",face='bold'),
                           axis.title = element_text(size=12,color="black",face='bold'),
                           axis.text = element_text(size=12,color="black",face='bold'),
                           axis.text.x = element_text(size=12,color="black",face='bold'),
                           axis.text.y = element_text(size=12,color="black",face='bold'),
                           legend.text = element_text(size=12,color="black",face='bold'),
                           legend.title = element_text(size=12,color="black",face='bold'),
                           title = element_text(size=12,color="black",face='bold'))
ggsave(filename = paste0(temp_out,"ASV_share_ra_by_family_vaginal_source_rf.pdf"), device = cairo_pdf, 
           plot = g, width = 9, height = 5, units = "in", dpi = 300)
```

## heat map based on shared ASV
```{r}
otu <- otu_table(phylo_decontam_clean)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()
dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>% arrange(FamilyNo, Generation)
shared_lst_no_rf <- lapply(Family_id, function(fid){
  wdat <- dat %>% filter(FamilyNo == fid)
  ft <- otu %>% dplyr::select(id, wdat$SampleID) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  Mom_id <- wdat %>% filter(Generation == "F0") %>% pull(SampleID)
  ft_mom <- ft[pmatch(Mom_id,row.names(ft)),]
  Pup_id <- wdat %>% filter(Generation == "F1") %>% pull(SampleID)
  res <- lapply(Pup_id, function(pup){
    ft_pup <- ft[pmatch(pup,row.names(ft)),]
    mom_asv <- which(ft_mom>0) %>% names()
    pup_asv <- which(ft_pup>0) %>% names()
    shared_asv <- intersect(mom_asv, pup_asv)
    shared_asv
  })
  res %>% unlist %>% unique()
})
names(shared_lst_no_rf) <- Family_id

otu <- otu_table(phylo_decontam_clean_rf)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()
dat <- wmeta %>% filter(FamilyNo %in% Family_id) %>% arrange(FamilyNo, Generation)
shared_lst_rf <- lapply(Family_id, function(fid){
  wdat <- dat %>% filter(FamilyNo == fid)
  ft <- otu %>% dplyr::select(id, wdat$SampleID) %>% column_to_rownames(var = "id") %>% t()
  ft <- ft[,colSums(ft)>0]
  Mom_id <- wdat %>% filter(Generation == "F0") %>% pull(SampleID)
  ft_mom <- ft[pmatch(Mom_id,row.names(ft)),]
  Pup_id <- wdat %>% filter(Generation == "F1") %>% pull(SampleID)
  res <- lapply(Pup_id, function(pup){
    ft_pup <- ft[pmatch(pup,row.names(ft)),]
    mom_asv <- which(ft_mom>1) %>% names()
    pup_asv <- which(ft_pup>1) %>% names()
    shared_asv <- intersect(mom_asv, pup_asv)
    shared_asv
  })
  res %>% unlist %>% unique()
})
names(shared_lst_rf) <- Family_id

library(ComplexHeatmap)
mat_col_fun = circlize::colorRamp2(c(-4, -3.3, -2.6, -1.9, -1.2, -0.5), c("#253494","#2c7fb8","#41b6c4", "#7fcdbb", "#c7e9b4", "#ffffcc"))
col_fun = circlize::colorRamp2(c(-2, 0, 2), c("green", "white", "red"))
taxa_new = taxa %>% mutate(taxa_long = str_c(Family,Genus,Species,ASV, sep = ";"))
```

### no rf
```{r}
otu <- otu_table(phylo_decontam_clean)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

ft <- otu %>% dplyr::select(id, wmeta$SampleID) %>% column_to_rownames(var = "id")
ft <- sweep(ft, 2, colSums(ft), "/")

#order samples
meta <- wmeta %>% mutate(Mom_or_Pup = ifelse(Generation == "F0", "Mom", "Pup"),
                         TP = ifelse(is.na(Age_group),"At birth", as.character(Age_group)),
                         TP = factor(TP, levels = c("At birth", "Week3",  "Week6", "Week9-14", "Week18"))) %>%
#  filter(FamilyNo %in% Family_id, Seq_count > depth) %>% arrange(TP)
  filter(FamilyNo %in% Family_id) %>% arrange(TP)

picked_taxa <- shared_lst_no_rf %>% unlist() %>% unique()
wft <- ft[picked_taxa, meta %>% pull(SampleID)]
#top 50
wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]
wft = log10(wft + 0.00001)
wft = wft %>% rownames_to_column(var="id") %>% 
  left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Mom_or_Pup, TP)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Mom_or_Pup, TP), 
                              col = list(Mom_or_Pup = c("Pup"="#7fc97f","Mom"="#beaed4"),
                                         TP = c(Age_color, "At birth"="#feedde")),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("", "Age_group"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            column_split = df_col$Mom_or_Pup,
            column_gap = unit(6, "mm"),
            column_title_gp = gpar(fontsize = 14, fontface = "bold"),
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(4, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("At birth","Week3", "Week6", "Week9-14", "Week18"), title = "Age_group", 
              legend_gp = gpar(fill = c("At birth"="#feedde", Age_color)),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("Mom","Pup"), title = "", 
              legend_gp = gpar(fill = c("Pup"="#7fc97f","Mom"="#beaed4")),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"sharing_asv_top50_all_no_rf.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

for (fid in Family_id){
  picked_taxa <- shared_lst_no_rf[[fid]]
  meta_fid <- meta %>% filter(FamilyNo == fid)
  wft <- ft[picked_taxa, meta_fid %>% pull(SampleID)]
#top 50
  if(nrow(wft)>50){wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]}
  wft = log10(wft + 0.00001)
  wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
    dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
  #rownames
  wft2 = wft %>% column_to_rownames(var = "taxa_long")
  df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
  #col annotation
  df_col = meta_fid %>% dplyr::select(Mom_or_Pup, TP)
  column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Mom_or_Pup, TP), 
                              col = list(Mom_or_Pup = c("Pup"="#7fc97f","Mom"="#beaed4"),
                                         TP = c("At birth"="#feedde",Age_color)),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("", "Age_group"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
  row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
  p1 = Heatmap(as.matrix(wft2), 
              col = mat_col_fun, 
              show_heatmap_legend = F,
              cluster_columns = F, show_column_names = F,
              column_split = df_col$Mom_or_Pup,
              column_gap = unit(4, "mm"),
              column_title_gp = gpar(fontsize = 14, fontface = "bold"),
              cluster_rows = row_dend,
              row_names_gp = gpar(fontsize = 14),
              row_labels = df_row_name_label,
              row_dend_width = unit(15, "mm"),
              row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
              top_annotation = column_ha,
              height = unit(8, "mm")*dim(wft)[1],
              width = unit(4, "mm")*dim(wft)[2]
              )
  p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"sharing_asv_",fid, "_no_rf.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()
}
```

### rf
```{r}
otu <- otu_table(phylo_decontam_clean_rf)  %>% as.data.frame() %>% rownames_to_column(var = "id")
Family_id_all <- meta_all %>% filter(Generation == "F0", SampleID %in% names(otu)[-1]) %>%
  pull(FamilyNo)
wmeta <- meta_all %>% filter(FamilyNo %in% Family_id_all, SampleID %in% names(otu)[-1])
Family_id <- which(table(wmeta$FamilyNo) > 1) %>% names()

ft <- otu %>% dplyr::select(id, wmeta$SampleID) %>% column_to_rownames(var = "id")
ft <- sweep(ft, 2, colSums(ft), "/")

#order samples
meta <- wmeta %>% mutate(Mom_or_Pup = ifelse(Generation == "F0", "Mom", "Pup"),
                         TP = ifelse(is.na(Age_group),"At birth", as.character(Age_group)),
                         TP = factor(TP, levels = c("At birth", "Week3",  "Week6", "Week9-14", "Week18"))) %>%
#  filter(FamilyNo %in% Family_id, Seq_count > depth) %>% arrange(TP)
  filter(FamilyNo %in% Family_id) %>% arrange(TP)

picked_taxa <- shared_lst_rf %>% unlist() %>% unique()
wft <- ft[picked_taxa, meta %>% pull(SampleID)]
#top 50
wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]
wft = log10(wft + 0.00001)
wft = wft %>% rownames_to_column(var="id") %>% 
  left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Mom_or_Pup, TP)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Mom_or_Pup, TP), 
                              col = list(Mom_or_Pup = c("Pup"="#7fc97f","Mom"="#beaed4"),
                                         TP = c(Age_color, "At birth"="#feedde")),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("", "Age_group"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            column_split = df_col$Mom_or_Pup,
            column_gap = unit(6, "mm"),
            column_title_gp = gpar(fontsize = 14, fontface = "bold"),
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(4, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("At birth","Week3", "Week6", "Week9-14", "Week18"), title = "Age_group", 
              legend_gp = gpar(fill = c("At birth"="#feedde", Age_color)),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("Mom","Pup"), title = "", 
              legend_gp = gpar(fill = c("Pup"="#7fc97f","Mom"="#beaed4")),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"sharing_asv_top50_all_rf.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

for (fid in Family_id){
  picked_taxa <- shared_lst_rf[[fid]]
  meta_fid <- meta %>% filter(FamilyNo == fid)
  wft <- ft[picked_taxa, meta_fid %>% pull(SampleID)]
#top 50
  if(nrow(wft)>50){wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]}
  wft = log10(wft + 0.00001)
  wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
    dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
  #rownames
  wft2 = wft %>% column_to_rownames(var = "taxa_long")
  df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
  #col annotation
  df_col = meta_fid %>% dplyr::select(Mom_or_Pup, TP)
  column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Mom_or_Pup, TP), 
                              col = list(Mom_or_Pup = c("Pup"="#7fc97f","Mom"="#beaed4"),
                                         TP = c("At birth"="#feedde",Age_color)),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("", "Age_group"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
  row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
  p1 = Heatmap(as.matrix(wft2), 
              col = mat_col_fun, 
              show_heatmap_legend = F,
              cluster_columns = F, show_column_names = F,
              column_split = df_col$Mom_or_Pup,
              column_gap = unit(4, "mm"),
              column_title_gp = gpar(fontsize = 14, fontface = "bold"),
              cluster_rows = row_dend,
              row_names_gp = gpar(fontsize = 14),
              row_labels = df_row_name_label,
              row_dend_width = unit(15, "mm"),
              row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
              top_annotation = column_ha,
              height = unit(8, "mm")*dim(wft)[1],
              width = unit(4, "mm")*dim(wft)[2]
              )
  p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
  pdf(file = paste0(temp_out,"sharing_asv_",fid, "_rf.pdf"), 
      width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
      height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
  draw(p)
  dev.off()
}
```
