==> code/06_taxa.Rmd <==
---
title: "Taxa"
author: "Haipeng Sun"
date: "7/21/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## library
```{r message=FALSE, warning=FALSE, include=FALSE}
library(phyloseq)
library(Hmisc)
library(agricolae)
library(grid)
library(gridExtra)
library(gtable)
library(ggrepel)
library(phyloseq)
library(usedist)
library(purrr)
library(furrr)
library(rlist)
library(ggpubr)
library(tidyverse)
```

## load data
```{r}
load(file = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/data/BW_and_phylo.RData")
temp_out = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/output/taxa/"
```

## masslin2
### function
```{r message=FALSE, warning=FALSE, include=FALSE}
library(Maaslin2)
### function wrap for all taxa levels
#in put metadata, sample id column name, phyloseq object
do_maaslin2 <- function(meta, sample_id, phylo_rf, fixed, random, ref, output, 
                       min_prevalence = 0.1, min_abundance = 0.0, max_significance = 0.1,
                       normalization = "TSS", transform = "LOG",analysis_method = "LM",
                       plot_heatmap	= FALSE, plot_scatter = FALSE, cores = 1){
  ft <- otu_table(phylo_rf@otu_table) %>% as.matrix() %>% as.data.frame() %>% rownames_to_column(var = "id")
  wft <- ft %>% dplyr::select(id,  meta %>% pull(sample_id)) %>% column_to_rownames(var = "id") %>% t()
  wmeta <- meta %>% column_to_rownames(var = sample_id)
  taxrank <- "ASV"
  fit <- Maaslin2(
    input_data = wft, 
    input_metadata = wmeta, 
    normalization = normalization,
    transform = transform,
    analysis_method = analysis_method,
    output = paste0(output,"_",taxrank),
    fixed_effects = fixed,
    reference = ref,
    random_effects = random,
    min_prevalence = min_prevalence,
    min_abundance = min_abundance,
    max_significance = max_significance,
    plot_heatmap = plot_heatmap,
    plot_scatter = plot_scatter,
    cores = cores
  )
  for (taxrank in c("Species","Genus","Family","Order","Class","Phylum")){
  phylo_glom <- tax_glom(phylo_rf, taxrank = taxrank, NArm = F)
  ft_glom <- otu_table(phylo_glom@otu_table) %>% as.matrix() %>% as.data.frame() %>% rownames_to_column(var = "id")
  wft <- ft_glom %>% dplyr::select(id, meta %>% pull(sample_id)) %>% column_to_rownames(var = "id") %>% t()
  fit <- Maaslin2(
      input_data = wft, 
      input_metadata = wmeta, 
      normalization = normalization,
      transform = transform,
      analysis_method = analysis_method,
      output = paste0(output,"_",taxrank),
      fixed_effects = fixed,
      reference = ref,
      random_effects = random,
      min_prevalence = min_prevalence,
      min_abundance = min_abundance,
      max_significance = max_significance,
      plot_heatmap = plot_heatmap,
      plot_scatter = plot_scatter,
      cores = cores
      )
  }
}

summarise_maaslin2 <- function(path, group, taxa){
  res <- matrix(data = NA, nrow = 1, ncol = 20)
  res <- as.data.frame(res)
  names(res) <-  c("feature","metadata","value","coef","stderr","N","N.not.0","pval","qval","Group",
                  "id","Kingdom","Phylum","Class","Order","Family","Genus","Species","ASV", "level")
  for (taxrank in c("ASV", "Species", "Genus", "Family", "Order", "Class", "Phylum")){
    a <- read.table(file = paste0(path ,"_", taxrank, "/significant_results.tsv"),
                  header = T, sep = "\t", stringsAsFactors = F)
    if (nrow(a) == 0) {next}
    a <- a %>% mutate(Group = group, id = str_remove(feature,"^X"),feature = str_remove(feature,"^X")) %>% left_join(taxa, by = "id") %>% mutate(level = taxrank)
    res <- rbind(res,a)
  }
  res <- res %>% filter(!is.na(feature))
  res
}
```

### compare
```{r message=FALSE, warning=FALSE, include=FALSE}
output <- paste0(temp_out, "maaslin2/")
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names

meta <- meta %>%  filter(Generation == "F1") %>% mutate(Age_group = as.character(Age_group))
#### include all Age_group
fit <- do_maaslin2(meta = meta, sample_id = "SampleID", phylo_rf = phylo_decontam_clean_rf, 
                  fixed = c("BirthMode", "Sex", "Age_group"), 
                  random = c("MouseID"), 
                  ref = c("BirthMode,VF", "Sex,F", "Age_group,Week3"), 
                  output = paste0(output,"All_Age_group"),
                  cores = 4
                  )
res <- summarise_maaslin2(path = paste0(output,"All_Age_group"), group = "compare_all", taxa)
write.csv(res, file = paste0(output,"All_Age_group_summary.csv"),quote = F, row.names = F)
#### by time points
for (tp in meta$Age_group %>% unique()){
  wmeta <- meta %>% filter(Age_group == tp)
  fit <- do_maaslin2(meta = wmeta, sample_id = "SampleID", phylo_rf = phylo_decontam_clean_rf, 
                  fixed = c("BirthMode", "Sex"), 
                  random = NULL, 
                  ref = c("BirthMode,VF", "Sex,F"), 
                  output = paste0(output,tp),
                  cores = 4
                  )
  res <- summarise_maaslin2(path = paste0(output,tp), group = tp, taxa)
  write.csv(res, file = paste0(output, tp, "_summary.csv"),quote = F, row.names = F)  
} 
```

## ancom
### function
```{r include=FALSE}
library(ANCOMBC)
### function wrap
do_ancombc <- function(phyloseq, formula = "",
                      p_adj_method = "BH", prv_cut = 0.1, lib_cut = 0,
                      group = NULL, struc_zero = FALSE, neg_lb = FALSE,
                      tol = 1e-05, max_iter = 100, conserve = FALSE, 
                      alpha = 0.05, global = FALSE, 
                      taxa = taxa){
  taxrank = "ASV"
  outbc <- ancombc(phyloseq = phyloseq, 
                   formula = formula,
                   p_adj_method = p_adj_method, prv_cut = prv_cut, lib_cut = lib_cut, 
                   group = group, struc_zero = struc_zero, neg_lb = neg_lb,
                   tol = tol, max_iter = max_iter, conserve = conserve, 
                   alpha = alpha, global = global)
  res1 <- as.data.frame(outbc$res) %>% rownames_to_column(var = "id") %>%
    left_join(taxa, by = "id") %>% mutate(level = taxrank)
  
  res_lst <- lapply(c("Species","Genus","Family","Order","Class","Phylum"), function(taxrank){
    phylo_glom = tax_glom(phyloseq, taxrank = taxrank, NArm = F)
    outbc <- ancombc(phyloseq = phylo_glom, 
                     formula = formula,
                     p_adj_method = p_adj_method, prv_cut = prv_cut, lib_cut = lib_cut, 
                     group = group, struc_zero = struc_zero, neg_lb = neg_lb,
                     tol = tol, max_iter = max_iter, conserve = conserve, 
                     alpha = alpha, global = global)   
    res <- as.data.frame(outbc$res) %>% rownames_to_column(var = "id") %>%
      left_join(taxa, by = "id") %>% mutate(level = taxrank)    
    res
  })
    res2 = do.call(rbind, res_lst)
    res <- rbind(res1, res2)
    res
}

```

### compare
```{r include=FALSE}
output <- paste0(temp_out, "ancom/")
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names

meta = meta %>%  filter(Generation == "F1") %>% mutate(Age_group = as.character(Age_group))
phylo_wk <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
new_sam_data <- sample_data(meta %>% column_to_rownames(var = "SampleID"))
sample_data(phylo_wk) <- new_sam_data

#### include all Age_group
res <- do_ancombc(phyloseq = phylo_wk,
                  formula = "BirthMode+Sex+Age_group", 
                  taxa = taxa)
write.csv(res, file = paste0(output, "All_Age_group_summary.csv"), 
          quote = F, row.names = F)

#### by time points
for (tp in meta$Age_group %>% unique()){
  wmeta = meta %>% filter(Age_group == tp)
  phylo_wk_tp <- prune_samples(wmeta$SampleID, phylo_wk)
  res <- do_ancombc(phyloseq = phylo_wk_tp,
                    formula = "BirthMode+Sex", 
                    taxa = taxa)
  write.csv(res, file = paste0(output, tp, "_summary.csv"),quote = F, row.names = F)  
} 

```

## compare different methods
```{r}
picked_feature <- list()
for (ll in c("ASV","Species","Genus","Family","Order","Class","Phylum")){
  picked_feature[[ll]][["Age_group"]][["All_Age_group"]][["maaslin2"]] <- read.csv(paste0(temp_out,"maaslin2/", "All_Age_group_summary.csv")) %>% filter(metadata == "Age_group", level == ll) %>% pull(feature)
  picked_feature[[ll]][["Age_group"]][["All_Age_group"]][["ancombc"]] <- read.csv(paste0(temp_out,"ancom/", "All_Age_group_summary.csv")) %>% filter(diff_abn.Age_groupWeek3|diff_abn.Age_groupWeek6|diff_abn.Age_groupWeek9.14) %>% filter(level == ll) %>% pull(id)
  for(tp in c("All_Age_group", "Week3", "Week6", "Week9-14", "Week18")){
    picked_feature[[ll]][["BrithMode"]][[tp]][["maaslin2"]] <- read.csv(paste0(temp_out,"maaslin2/",tp, "_summary.csv")) %>% 
      filter(metadata == "BirthMode", level == ll) %>% pull(feature)
    picked_feature[[ll]][["BrithMode"]][[tp]][["ancombc"]] <- read.csv(paste0(temp_out,"ancom/",tp, "_summary.csv")) %>% 
      filter(diff_abn.BirthModeCSR | diff_abn.BirthModeVF) %>% filter(level == ll) %>% pull(id)    
    picked_feature[[ll]][["Sex"]][[tp]][["maaslin2"]] <- read.csv(paste0(temp_out,"maaslin2/",tp, "_summary.csv")) %>% 
      filter(metadata == "Sex", level == ll) %>% pull(feature)
    picked_feature[[ll]][["Sex"]][["ancombc"]] <- read.csv(paste0(temp_out,"ancom/",tp, "_summary.csv")) %>% 
      filter(diff_abn.SexM) %>% filter(level == ll) %>% pull(id)    
  }
}

#save(picked_feature, file = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/data/picked_feature.Rdata")
load(file = "~/Dropbox/rutgers/20220722_IFNH_CSR_new/data/picked_feature.Rdata")

map_lst <-list(taxrank = c("ASV","Species","Genus","Family","Order","Class","Phylum"), 
     Age_group = c("All_Age_group", "Week3", "Week6", "Week9-14", "Week18")) %>% cross()
res <- map_dfr(map_lst, function(x){
  m_lst <- picked_feature[[x$taxrank]][["BrithMode"]][[x$Age_group]]$maaslin2
  a_lst <- picked_feature[[x$taxrank]][["BrithMode"]][[x$Age_group]]$ancombc
  res <- data.frame(level = x$taxrank, Age_group = x$Age_group,
                    maaslin2 = length(m_lst),
                    ancombc = length(a_lst),
                    common = length(intersect(m_lst, a_lst)))
  res
})
write.csv(res %>% arrange(level, Age_group),
          file = paste0(temp_out,"feature_number_picked_birthmode.csv"),
          quote = F, row.names = F)
```

## heatmaps
```{r message=FALSE, warning=FALSE, include=FALSE}
library(ComplexHeatmap)

mat_col_fun = circlize::colorRamp2(c(-4, -3.3, -2.6, -1.9, -1.2, -0.5), c("#253494","#2c7fb8","#41b6c4", "#7fcdbb", "#c7e9b4", "#ffffcc"))
col_fun = circlize::colorRamp2(c(-2, 0, 2), c("green", "white", "red"))
taxa_new = taxa %>% mutate(taxa_long = str_c(Family,Genus,Species, sep = ";"))
```

### use ancombc species level result to plot time point
```{r }
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names
meta <- meta %>% filter(Generation == "F1")
phylo_wk <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
phylo_s <- tax_glom(phylo_wk, taxrank = "Species", NArm = F)
ft <- otu_table(phylo_s@otu_table)  %>% as.data.frame() 

#order samples
meta <- meta %>% arrange(Age_group, BirthMode, Sex)
picked_taxa <- picked_feature$Species$Age_group$All_Age_group$ancombc
wft <- ft[picked_taxa, meta %>% pull(SampleID)]
#top 50 
wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]
wft = log10((wft + 0.5)/colSums(ft[,-1])[1])
wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Age_group, BirthMode, Sex)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Age_group, BirthMode), 
                              col = list(Age_group = Age_color,
                                         BirthMode = BirthMode_color),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("Age group", "BirthMode"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            #name = "Relative Abundance",
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            #column_split = df_col$SamplingAge_group,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("Week3", "Week6", "Week9-14", "Week18"), title = "Age group", 
              legend_gp = gpar(fill = Age_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("CS","CSR","VF"), title = "BirthMode", 
              legend_gp = gpar(fill = BirthMode_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"ancom_Age_group_s_supervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

p2 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = T, show_column_names = F,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"ancom_Age_group_s_unsupervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

```

### use ancombc species level result to plot birthmode
```{r }
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names
meta <- meta %>% filter(Generation == "F1")
phylo_wk <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
phylo_s <- tax_glom(phylo_wk, taxrank = "Species", NArm = F)
ft <- otu_table(phylo_s@otu_table)  %>% as.data.frame() 

#order samples
meta <- meta %>% arrange(BirthMode,Age_group,  Sex)
picked_taxa <- picked_feature$Species$BrithMode$All_Age_group$ancombc
wft <- ft[picked_taxa, meta %>% pull(SampleID)]

wft = log10((wft + 0.5)/colSums(ft[,-1])[1])
wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Age_group, BirthMode, Sex)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Age_group,BirthMode), 
                              col = list(Age_group = Age_color,
                                         BirthMode = BirthMode_color),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("Age Group", "BirthMode"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            #name = "Relative Abundance",
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            #column_split = df_col$SamplingAge_group,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("Week3", "Week6", "Week9-14", "Week18"), title = "Age group", 
              legend_gp = gpar(fill = Age_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("CS","CSR","VF"), title = "BirthMode", 
              legend_gp = gpar(fill = BirthMode_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"ancom_birthmode_s_supervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

p2 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = T, show_column_names = F,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"ancom_birthmode_s_unsupervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

```

### use maaslin2 species level result to plot time point
```{r }
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names
meta <- meta %>% filter(Generation == "F1")
phylo_wk <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
phylo_s <- tax_glom(phylo_wk, taxrank = "Species", NArm = F)
ft <- otu_table(phylo_s@otu_table)  %>% as.data.frame() 

#order samples
meta <- meta %>% arrange(Age_group, BirthMode, Sex)
picked_taxa <- picked_feature$Species$Age_group$All_Age_group$maaslin2 %>% unique()
wft <- ft[picked_taxa, meta %>% pull(SampleID)]
#top 50 
wft <- wft[(rowSums(wft) %>% sort(.,decreasing = T))[1:50] %>% names(),]
wft = log10((wft + 0.5)/colSums(ft[,-1])[1])
wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Age_group, BirthMode, Sex)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select(Age_group, BirthMode), 
                              col = list(Age_group = Age_color,
                                         BirthMode = BirthMode_color),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("Age group", "BirthMode"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            #name = "Relative Abundance",
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            #column_split = df_col$SamplingAge_group,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("Week3", "Week6", "Week9-14", "Week18"), title = "Age group", 
              legend_gp = gpar(fill = Age_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("CS","CSR","VF"), title = "BirthMode", 
              legend_gp = gpar(fill = BirthMode_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"maaslin2_Age_group_s_supervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

p2 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = T, show_column_names = F,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"maaslin2_Age_group_s_unsupervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

```

### use maaslin2 species level result to plot birthmode
```{r }
meta <- phylo_decontam_clean_rf@sam_data@.Data %>% as.data.frame()
names(meta) <- phylo_decontam_clean_rf@sam_data@names
meta$SampleID <- phylo_decontam_clean_rf@sam_data@row.names
meta <- meta %>% filter(Generation == "F1")
phylo_wk <- prune_samples(meta$SampleID, phylo_decontam_clean_rf)
phylo_s <- tax_glom(phylo_wk, taxrank = "Species", NArm = F)
ft <- otu_table(phylo_s@otu_table)  %>% as.data.frame() 

#order samples
meta <- meta %>% arrange(BirthMode,Age_group,  Sex)
picked_taxa <- picked_feature$Species$BrithMode$All_Age_group$maaslin2 %>% unique()
wft <- ft[picked_taxa, meta %>% pull(SampleID)]

wft = log10((wft + 0.5)/colSums(ft[,-1])[1])
wft = wft %>% rownames_to_column(var="id") %>% left_join(taxa_new %>% dplyr::select(id, taxa_long), by = "id") %>% 
  dplyr::select(-id) %>% dplyr::select(taxa_long, everything())
#rownames
wft2 = wft %>% column_to_rownames(var = "taxa_long")
df_row_name_label = structure(rownames(wft2), names = rownames(wft2))
#col annotation
df_col = meta %>% dplyr::select(Age_group, BirthMode, Sex)
column_ha = HeatmapAnnotation(df = df_col %>% dplyr::select( Age_group,BirthMode), 
                              col = list(Age_group = Age_color,
                                         BirthMode = BirthMode_color),
                              simple_anno_size = unit(6, "mm"), 
                              annotation_label = c("Age Group", "BirthMode"),
                              annotation_name_gp = gpar(fontsize = 14,fontface = "bold"), 
                              annotation_name_side = "left",
                              show_legend = F)
#row dend order
row_dend = as.dendrogram(hclust(dist(wft2)))
#heatmap
p1 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            #name = "Relative Abundance",
            show_heatmap_legend = F,
            cluster_columns = F, show_column_names = F,
            #column_split = df_col$SamplingAge_group,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
lgd1 = Legend(labels = c("Week3", "Week6", "Week9-14", "Week18"), title = "Age group", 
              legend_gp = gpar(fill = Age_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd2 = Legend(labels = c("CS","CSR","VF"), title = "BirthMode", 
              legend_gp = gpar(fill = BirthMode_color),
              title_position = "topleft", 
              labels_gp = gpar(fontsize = 14, fontface = "bold"),
              title_gp = gpar(fontsize = 14, fontface = "bold"))
lgd3 = Legend(col_fun = mat_col_fun, title = "Relative abundance", 
              at = c(-4, -3.3, -2.6, -1.9, -1.2, -0.5),                 
              labels = c("0.01%", "0.05%", "0.25%", "1.25%", "6.25%", "31.25%"), 
              title_position = "topleft",
              labels_gp = gpar(fontsize = 14, fontface = "bold"), 
              title_gp = gpar(fontsize = 14, fontface = "bold"))
pd = packLegend(list = list(lgd1, lgd2, lgd3), direction = "vertical")

p = draw(p1, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"maaslin2_birthmode_s_supervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

p2 = Heatmap(as.matrix(wft2), 
            col = mat_col_fun, 
            show_heatmap_legend = F,
            cluster_columns = T, show_column_names = F,
            cluster_rows = row_dend,
            row_names_gp = gpar(fontsize = 14),
            row_labels = df_row_name_label,
            row_dend_width = unit(15, "mm"),
            row_names_max_width = max_text_width(df_row_name_label, gp = gpar(fontsize = 14)),
            top_annotation = column_ha,
            height = unit(8, "mm")*dim(wft)[1],
            width = unit(1, "mm")*dim(wft)[2]
            )
p = draw(p2, annotation_legend_side = "left", annotation_legend_list = pd)
pdf(file = paste0(temp_out,"maaslin2_birthmode_s_unsupervised.pdf"), 
    width = ComplexHeatmap:::width(p) %>% as.numeric()/25.4, 
    height = ComplexHeatmap:::height(p) %>% as.numeric()/25.4, useDingbats = F)
draw(p)
dev.off()

```
