==> code/08_fat_tissue_PCR.Rmd <==
---
title: "Fat tissue"
author: "Haipeng Sun"
date: "7/25/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## library
```{r message=FALSE, warning=FALSE, include=FALSE}
library(grid)
library(gridExtra)
library(gtable)
library(ggrepel)
library(rlist)
library(ggpubr)
library(readxl)
library(rstatix)
library(ggpattern)
library(patchwork)
library(tidyverse)
```

## load data
```{r}
path_rec = "~/Dropbox/Rutgers/20220722_IFNH_CSR_new/"
load(file = paste0(path_rec,"data/general_data.Rdata"))
load(file = paste0(path_rec,"data/BW_responder.RData"))
```

## fat
### load fat PCR result
```{r}
BAT_file = paste0(path_rec,"data/ExperimentRecords/Result_from_Harini/BAT qPCR_New spls_11112021.xlsx")
iWAT_file = paste0(path_rec,"data/ExperimentRecords/Result_from_Harini/iWAT qPCR_New spls_11112021.xlsx")

df <- list()
for (ff in c("BAT","iWAT")){
  for (gene in c("UCP1","Cidea","Pgc 1a","Prdm 16", "Dio2","Sirt 1")){
    a <- read_excel(get(paste0(ff, "_file")), sheet = paste0(ff, "_", gene), range = "B2:J28")
    a <-  a %>% filter(!is.na(Gene)) %>% 
      transmute(FamilyID = `Animal Number`, BirthMode = str_remove(`Animal Number`,"\\d+"),
                Gene, Ct = CT, `18 s`, Delta_Ct = `Delta CT`, DD_Ct = `ddCt relative to WT`,
                FC = `Fold Change`, Tissue = ff)
    df <- list.append(df, a)
  }
}

df2 <- do.call(rbind, df) %>% 
  filter(!is.na(FC))

df_sum <-  df2 %>% group_by(Tissue, Gene, BirthMode) %>% 
  summarise(FC_mean = mean(FC), FC_sd = sd(FC), N = n()) %>% ungroup() %>% 
  mutate(FC_se = FC_sd / sqrt(N))
  

norm_test <- df2 %>% group_by(Tissue, Gene) %>% 
  shapiro_test(FC)

stat.test <- df2 %>% group_by(Tissue, Gene) %>% 
  wilcox_test(FC ~ BirthMode) %>% 
  adjust_pvalue(method = "fdr") %>% 
  add_significance("p.adj") %>% 
  add_xy_position(x = "BirthMode", scales = "free", fun = "mean_sd")

```

### plot
```{r}
temp_out = paste0(path_rec,"output/fat/")

g <- ggplot(df2, aes(Gene, FC, color = BirthMode)) + 
  geom_point(position = position_dodge(width = 0.9)) + 
  geom_col(data = df_sum, aes(Gene, FC_mean, color = BirthMode, fill = BirthMode),
           position = position_dodge(), alpha = 0.5) +
  facet_grid(Tissue ~ ., scales = "free_y")

g_list <- list()

for (ff in c("BAT","iWAT")){
  for (gene in c("UCP1","Cidea","Pgc 1a","Prdm 16", "Dio2","Sirt 1")){
    wdat <- df2 %>% filter(Tissue == ff, Gene == gene)
    wdat_sum <- df_sum %>% filter(Tissue == ff, Gene == gene) 
    wstat.test <- stat.test %>% filter(Tissue == ff, Gene == gene)
    g_list[[ff]][[gene]]  <- ggplot(wdat, aes(BirthMode, FC, color = BirthMode)) + 
      geom_point() + 
      geom_col(data = wdat_sum, aes(BirthMode, FC_mean, fill = BirthMode), 
               alpha = 0.7, width = 0.3) + 
      geom_errorbar(data = wdat_sum, aes(BirthMode, FC_mean, ymin = FC_mean - FC_se, ymax = FC_mean + FC_se),
                    width = 0.1) +
      stat_pvalue_manual(wstat.test, label = "p.adj.signif", tip.length = 0, hide.ns = T, size = 12/.pt) + 
      scale_color_manual(values = BirthMode_color) +
      scale_y_continuous(expand = c(0, 0), limits = c(0, max(wdat$FC)*1.05)) +
      labs(x = "", y = "Relative expression\n(Normalized to 18S RNA)", title = gene) + 
      theme_classic() + theme(aspect.ratio = 1,
                              panel.background = element_rect(fill = NA),
                              plot.title = element_text(size=14,color="black",face='bold',hjust=0.5),
                              axis.title.y = element_text(size=12,color="black",face='bold'),
                              axis.text.x = element_text(size=12,color="black",face='bold'),
                              axis.text.y = element_text(size=12,color="black",face='bold'),
                              legend.text = element_text(size=12,color="black",face='bold'),
                              legend.title = element_text(size=12,color="black",face='bold')) 
  }
}

for (ff in c("BAT","iWAT")){
  p0 <- g_list[[ff]][["Cidea"]]
  for (gene in c("Dio2","Pgc 1a","Prdm 16","Sirt 1","UCP1")){
    p0 <- p0 + g_list[[ff]][[gene]]
  }
  p0 <- p0 + plot_layout(guides = "collect", ncol = 6) + 
    plot_annotation(title = "", tag_levels = 'a') & theme(plot.tag = element_text(size = 12, face = "bold"), 
                                                          legend.position = "right")
  ggsave(filename = paste0(temp_out,ff, "_browning_marker_by_birthmode_QPCR.pdf"), device = cairo_pdf, 
         plot = p0, width = 15, height = 4, units = "in", dpi = 300)
  
}

```

## CLAMS
### load data
```{r}
CLAMS_file = paste0(path_rec,"data/ExperimentRecords/Result_from_Troy/DB IFNH CLAMS Analysis.xlsx")
Mapping_file = paste0(path_rec,"data/ExperimentRecords/Result_from_Troy/BehaviorAnimalMetadata-20190124-JW.xlsx")

dat <- lapply(excel_sheets(CLAMS_file),function(tt){
  a <- read_excel(CLAMS_file, sheet = tt, range = "A3:Y42", col_names = F)
  names(a) <- c("MouseID", c(0:23))
  a <- a %>% mutate(MouseID = str_to_upper(MouseID)) %>% 
    mutate(MouseID = ifelse(MouseID == "FO59.P8", "F059.P8", MouseID))
  aa <- a %>% pivot_longer(cols = -MouseID, values_to = "Value", names_to = "Time") %>% 
    filter(!is.na(Value)) %>% 
    mutate(Test = tt)
  aa
}) %>% do.call(rbind,.) 

map <- read_excel(Mapping_file) 

dat <- dat %>% 
  left_join(map, by = "MouseID") %>% 
  mutate(Time = as.integer(Time)) %>% 
  mutate(Day_or_Night = ifelse(Time>6&Time<19, "Day","Night"))

#hourly
dat_sum <-  dat %>% group_by(Time, Test, BirthMode, Sex) %>% 
  summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% ungroup() %>% 
  mutate(Value_se = Value_sd / sqrt(N))
  
norm_test <- dat %>% group_by(Time, Test) %>% 
  shapiro_test(Value) %>% 
  mutate(P2 = round(p, digits = 3)) %>% 
  mutate(norm = ifelse(P2>0.05, "Y","N")) %>% count(Test, norm) %>% 
  pivot_wider(Test, names_from = norm, values_from = n)

stat.test1 <- dat %>% group_by(Time, Test, Sex) %>% 
  t_test(Value ~ BirthMode) %>% 
  adjust_pvalue(method = "BH") %>% 
  add_significance("p") %>% 
  add_xy_position(x = "BirthMode", scales = "free", fun = "mean_sd")

stat.test2 <- stat.test1 %>% 
  mutate(grp = str_c(group1, group2, sep = "_")) %>% 
  mutate(sig_lab = case_when(p.adj.signif != "ns" & grp == "CS_VF" ~ "A",
                             p.adj.signif != "ns" & grp == "CSR_VF" ~ "B",
                             p.adj.signif != "ns" & grp == "CS_CSR" ~ "C",
                             p.adj.signif == "ns" ~ "")) %>%
  arrange(Test, Time, Sex, sig_lab) %>% 
  group_by(Test, Time, Sex) %>% 
  summarize(sig_lab1 = paste(sig_lab, collapse = "")) %>% ungroup()

stat.test3 <-  df_sum %>% mutate(Value_max = Value_mean + Value_se) %>% 
  group_by(Time, Test, Sex) %>% 
  summarise(Value_max = max(Value_max)) %>% ungroup() %>% 
  left_join(stat.test2, by = c("Test","Time","Sex"))

#day night
dat_dn <- dat %>% group_by(MouseID, Day_or_Night, Test) %>% 
  summarise(Value = mean(Value)) %>% ungroup() %>% 
  left_join(map, by = "MouseID")

dat_dn_sum <-  dat %>% group_by(Day_or_Night, Test, BirthMode, Sex) %>% 
  summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% ungroup() %>% 
  mutate(Value_se = Value_sd / sqrt(N))

stat.dn.test1 <- dat_dn %>% group_by(Day_or_Night, Test, Sex) %>% 
  t_test(Value ~ BirthMode) %>% 
  adjust_pvalue(method = "BH") %>% 
  add_significance("p") %>% 
  add_xy_position(x = "BirthMode", scales = "free", fun = "mean_sd")

```

### plot
```{r}
tt = "V.O2"
ss = "M"

wdat <- dat %>% filter(Test == tt)
wdat_sum <- wdat %>% group_by(Sex, Time, BirthMode) %>% 
  summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% 
  ungroup() %>% 
  mutate(Value_se = Value_sd/sqrt(N))
wstat <- stat.test3 %>% filter(Test == tt)

y_max <- max(wdat_sum$Value_mean + wdat_sum$Value_sd)*1.1
y_min <- 0

plot_hourly <- function(wdat_sum, wstat, ylab, tt, y_max, y_min, ss){
  wwdat_sum <- wdat_sum %>% filter(Sex == ss)
  wwstat <- wstat %>% filter(Sex == ss)
  g <- ggplot(wwdat_sum, aes(Time, Value_mean)) + 
    geom_point(aes(color = BirthMode)) + 
    geom_path(aes(group = BirthMode, color = BirthMode)) + 
    geom_errorbar(aes(ymin = Value_mean - Value_se, ymax = Value_mean + Value_se, color = BirthMode),
                  width = 0.1) + 
    geom_text(data = wwstat, aes(Time, Value_max, label = sig_lab1), vjust = 0) + 
    geom_segment(data = data.frame(Time = c(0,19), Value_mean = c(y_min, y_min), T2 = c(7,24)), 
                 aes(x=Time, xend=T2, y=Value_mean, yend=Value_mean),
                 linewidth = 2) + 
    scale_x_continuous(breaks = seq(0,23, by = 2)) + 
    scale_y_continuous(n.breaks = 6, limits = c(y_min, y_max)) + 
    labs(x = "Hour", y = ylab, title = paste(tt, ss)) + 
    theme_classic() + theme(aspect.ratio = 1,
                            panel.background = element_rect(fill = NA),
                            plot.title = element_text(size=14,color="black",face='bold',hjust=0.5),
                            axis.title = element_text(size=12,color="black",face='bold'),
                            axis.title.y = element_text(size=12,color="black",face='bold'),
                            axis.text.x = element_text(size=12,color="black",face='bold'),
                            axis.text.y = element_text(size=12,color="black",face='bold'),
                            legend.text = element_text(size=12,color="black",face='bold'),
                            legend.title = element_text(size=12,color="black",face='bold')) 
  g
}
g1 <- plot_hourly(wdat_sum, wstat, ylab, tt, y_max, y_min, ss)

wdat_dn <- dat_dn %>% filter(Test == tt)
wdat_dn_sum <- wdat_dn %>% group_by(Sex, Day_or_Night, BirthMode) %>% 
  summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% 
  ungroup() %>% 
  mutate(Value_se = Value_sd/sqrt(N))
wstat <- stat.dn.test1 %>% filter(Test == tt)

y_max <- max(wdat_dn_sum$Value_mean + wdat_dn_sum$Value_se)*1.1
y_min <- 0

plot_day_night <- function(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss){
  dat_ <- wdat_dn_sum %>% filter(Sex == ss)
  wwstat <- wstat %>% filter(Sex == ss)
  g <- ggplot(dat_, aes(BirthMode, Value_mean)) + 
    #geom_col(aes(fill = BirthMode), position = position_dodge2()) + 
    geom_col_pattern(aes(pattern = Day_or_Night, fill = BirthMode), 
                     position = position_dodge2(), color = "black",  
                     pattern_spacing = 0.01, pattern_density = 0.01, pattern_fill = "black") + 
    scale_pattern_manual(values=c('none', 'stripe'), guide = "none") + 
    geom_errorbar(aes(ymin = Value_mean, ymax = Value_mean + Value_se), 
                  position = position_dodge2(width = 0.5, padding = 0.5)) + 
    stat_pvalue_manual(data = wwstat, label = "p.adj.signif", hide.ns = TRUE, 
                       tip.length = 0) + 
    scale_color_manual(values = BirthMode_color) + 
    scale_x_discrete(labels = c("Day   Night","Day   Night","Day   Night")) + 
    scale_y_continuous(n.breaks = 6, limits = c(y_min, y_max), expand = c(0,0)) + 
    labs(x = "", y = ylab, title = paste(tt, ss)) + 
    guides(fill = guide_legend(override.aes = list(pattern = c("none", "none", "none")))) +
    theme_classic() + theme(aspect.ratio = 1,
                            panel.background = element_rect(fill = NA),
                            plot.title = element_text(size=14,color="black",face='bold',hjust=0.5),
                            axis.title = element_text(size=12,color="black",face='bold'),
                            axis.title.y = element_text(size=12,color="black",face='bold'),
                            axis.text.x = element_text(size=12,color="black",face='bold', hjust = 0.4),
                            axis.text.y = element_text(size=12,color="black",face='bold'),
                            legend.text = element_text(size=12,color="black",face='bold'),
                            legend.title = element_text(size=12,color="black",face='bold'))
  g
}
g2 <- plot_day_night(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss)

g_lst <- list()
for (tt in dat$Test %>% unique()){
  if(tt == "V.O2"){
    ylab = expression(bold(V.O[2] (ml/min/kg)))
  }else if(tt == "V.CO2"){
    ylab = expression(bold(V.CO[2] (ml/min/kg)))
  }else if(tt == "RER"){
    ylab = expression(bold(RER (VCO[2]/VO[2])))
  }else if(tt == "Heat"){
    ylab = "Heat (kCal/hr)"
  }else if(tt == "X total"){
    ylab = "X total (counts)"
  }else if(tt == "Y total"){
    ylab = "Y total (counts)"
  }else if(tt == "WHEEL"){
    ylab = "Wheel Counts"
  }else if(tt == "Drink"){
    ylab = "Water intake (g/h)"
  }else{
    ylab = tt
  }
  
  wdat <- dat %>% filter(Test == tt)
  wdat_sum <- wdat %>% group_by(Sex, Time, BirthMode) %>% 
    summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% 
    ungroup() %>% 
    mutate(Value_se = Value_sd/sqrt(N))
  wstat <- stat.test3 %>% filter(Test == tt)
  y_max <- max(wdat_sum$Value_mean + wdat_sum$Value_se)*1.1
  y_min <- 0
  if(tt == "RER"){y_min = 0.8}
  if(tt %in% c("X amb", "Y amb", "WHEEL")){
    y_max <- wdat_sum %>% filter(Sex == "M") %>% 
      mutate(Value_max = (Value_mean+Value_se)*1.1) %>% 
      pull(Value_max) %>% max()
    g_lst[[tt]][["hr"]][["M"]] <- plot_hourly(wdat_sum, wstat, ylab, tt, y_max, y_min, ss = "M")
    y_max <- wdat_sum %>% filter(Sex == "F") %>% 
      mutate(Value_max = (Value_mean+Value_se)*1.1) %>% 
      pull(Value_max) %>% max()
    g_lst[[tt]][["hr"]][["F"]] <- plot_hourly(wdat_sum, wstat, ylab, tt, y_max, y_min, ss = "F")
  }else{
    g_lst[[tt]][["hr"]][["M"]] <- plot_hourly(wdat_sum, wstat, ylab, tt, y_max, y_min, ss = "M")
    g_lst[[tt]][["hr"]][["F"]] <- plot_hourly(wdat_sum, wstat, ylab, tt, y_max, y_min, ss = "F")
  }
  
  wdat_dn <- dat_dn %>% filter(Test == tt)
  wdat_dn_sum <- wdat_dn %>% group_by(Sex, Day_or_Night, BirthMode) %>% 
    summarise(Value_mean = mean(Value), Value_sd = sd(Value), N = n()) %>% 
    ungroup() %>% 
    mutate(Value_se = Value_sd/sqrt(N))
  wstat <- stat.dn.test1 %>% filter(Test == tt)
  y_max <- max(wdat_dn_sum$Value_mean + wdat_dn_sum$Value_se)*1.1
  y_min <- 0
  if(tt %in% c("X amb", "Y amb", "WHEEL")){
    y_max <- wdat_dn_sum %>% filter(Sex == "M") %>% 
      mutate(Value_max = (Value_mean+Value_se)*1.1) %>% 
      pull(Value_max) %>% max()
    g_lst[[tt]][["dn"]][["M"]] <- plot_day_night(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss = "M")
    y_max <- wdat_dn_sum %>% filter(Sex == "F") %>% 
      mutate(Value_max = (Value_mean+Value_se)*1.1) %>% 
      pull(Value_max) %>% max()
    g_lst[[tt]][["dn"]][["F"]] <- plot_day_night(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss = "F")
  }else{
    g_lst[[tt]][["dn"]][["M"]] <- plot_day_night(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss = "M")
    g_lst[[tt]][["dn"]][["F"]] <- plot_day_night(wdat_dn_sum, wstat, ylab, tt, y_max, y_min, ss = "F")
  }  
  
  p <- g_lst[[tt]][["hr"]][["M"]] + g_lst[[tt]][["dn"]][["M"]] + 
    g_lst[[tt]][["hr"]][["F"]] + g_lst[[tt]][["dn"]][["F"]] + 
    plot_layout(guides = "collect", ncol = 2) + 
    plot_annotation(title = "", tag_levels = 'A') & theme(plot.tag = element_text(size = 12, face = "bold"), 
                                                          legend.position = "right")
  ggsave(filename = paste0(path_rec,"output/CLAMS/test_", tt,"_panel.pdf"), device = cairo_pdf, 
       plot = p, width = 10, height = 10, units = "in", dpi = 300)
}



```
## Glucose
### load data
```{r}
glu_1 = paste0(path_rec,"data/ExperimentRecords/Result_from_Troy/DB Behavior1a 1b GTT ITT.xlsx")
glu_2 = paste0(path_rec,"data/ExperimentRecords/Result_from_Troy/DB Behavior2a 2b GTT ITT.xlsx")

Mapping_file = paste0(path_rec,"data/ExperimentRecords/Result_from_Troy/BehaviorAnimalMetadata-20190124-JW.xlsx")

map <- read_excel(Mapping_file) 

dat1 <- lapply(excel_sheets(glu_1),function(tt){
  a <- read_excel(glu_1, sheet = tt, range = "B3:L19", col_names = T)
  a <- a %>% filter(!is.na(`ID#`)) %>% 
    transmute(MouseID = `ID#`, BW, Inj_vol = `Inj vol`,
              T0 = `time = 0`, T15 = `15`, T30 = `30`, T60 = `60`, T90 = `90`, T120 = `120`) %>% 
    mutate(MouseID = str_replace(MouseID, "f", "F")) %>% 
    mutate(MouseID = str_replace(MouseID, "p", "P")) %>% 
    mutate(Test = ifelse(str_detect(tt,"GTT"),"GTT","ITT"))
  a
}) %>% do.call(rbind,.) 

dat2 <- lapply(excel_sheets(glu_2),function(tt){
  a <- read_excel(glu_2, sheet = tt, range = "B3:L19", col_names = T)
  a <- a %>% filter(!is.na(`ID#`)) %>% 
    transmute(MouseID = `ID#`, BW, Inj_vol = `Inj vol`,
              T0 = `time = 0`, T15 = `15`, T30 = `30`, T60 = `60`, T90 = `90`, T120 = `120`) %>% 
    mutate(MouseID = str_replace(MouseID, "f", "F")) %>% 
    mutate(MouseID = str_replace(MouseID, "p", "P")) %>% 
    mutate(Test = ifelse(str_detect(tt,"GTT"),"GTT","ITT"))
  a
}) %>% do.call(rbind,.) 

dat <- rbind(dat1, dat2) %>% 
  pivot_longer(cols = c(T0, T15, T30, T60, T90, T120), names_to = "Time", values_to = "Value_text") %>%
  left_join(map, by = "MouseID")
# do I fill HI with max value
# dat %>% group_by(Test) %>% summarize(max(as.integer(Value_text), na.rm = T))
# 
dat <- dat %>%
  mutate(Value = case_when(Value_text == "HI" & Test == "GTT" ~ 750,
                            Value_text == "HI" & Test == "ITT" ~ 651,
                            TRUE ~ as.integer(Value_text)))
#dat <- dat %>% mutate(Value = as.integer(Value_text))
#*animal excluded due to "HI" time 0
dat <- dat %>% filter(MouseID != "F036.P7")

norm_test <- dat %>% group_by(Time, Test) %>% 
  shapiro_test(Value) %>% 
  mutate(Norm = ifelse(p>0.05, "Y", "N"))

stat.test1 <- dat %>% group_by(Time, Test, Sex) %>% 
  dunn_test(Value ~ BirthMode) %>% 
  adjust_pvalue(method = "BH") %>% 
  add_significance("p.adj") %>% 
  add_xy_position(x = "BirthMode", scales = "free", fun = "mean_sd")

dat_sum <- dat %>% group_by(Time, Test, Sex, BirthMode) %>% 
  summarise(Value_mean = mean(Value, na.rm = TRUE), Value_sd = sd(Value, na.rm = TRUE), N = sum(!is.na(Value))) %>% 
  ungroup() %>% 
  mutate(Value_se = Value_sd/sqrt(N))

## AUC
dat_w <- dat %>% pivot_wider(id_cols = c(MouseID, Test, Sex, BirthMode), 
                             names_from = Time, values_from = Value) %>% 
  mutate(AUC = 1/2*(T0+T30) + 1/2*(T30+T60) + 1/2*(T60+T90) + 1/2*(T90+T120))

```

### plot
```{r}

# tt in c("GTT","ITT")
# ss in c("M","F") 
tt = "GTT";ss = "M"
if(tt == "GTT"){}
wdat_sum <- dat_sum %>% filter(Test == tt, Sex == ss)
wdat_w <- dat_w %>% filter(Test == tt, Sex == ss)

plot_time_value <- function(wdat_sum, tt, ss){
  dat_ <- wdat_sum %>% 
    mutate(Time = as.integer(str_remove(Time, "^T"))) %>% 
    arrange(Time)
  ylab = "Glucose (mg/dL)"
  g <- ggplot(dat_, aes(x = Time, Value_mean, color = BirthMode)) +
    geom_point() + 
    geom_path(aes(group = BirthMode)) + 
    geom_errorbar(aes(ymin = Value_mean - Value_se, ymax = Value_mean + Value_se),
                  width = 0.9) + 
    scale_x_continuous(breaks = c(0,15,30,45,60,75,90,105,120)) + 
    scale_color_manual(values = BirthMode_color) + 
    labs(x = "Time min", y = ylab, title = paste(tt, ss)) +
    theme_classic() + theme(aspect.ratio = 1,
                            panel.background = element_rect(fill = NA),
                            plot.title = element_text(size=14,color="black",face='bold',hjust=0.5),
                            axis.title = element_text(size=12,color="black",face='bold'),
                            axis.text.x = element_text(size=12,color="black",face='bold'),
                            axis.text.y = element_text(size=12,color="black",face='bold'),
                            legend.text = element_text(size=12,color="black",face='bold'),
                            legend.title = element_text(size=12,color="black",face='bold'))
  g
}


plot_auc <- function(wdat_w, tt, ss){
  dat_ <- wdat_w %>% 
    group_by(BirthMode) %>% 
    summarise(Value_mean = mean(AUC), Value_sd = sd(AUC), N = n()) %>% 
    ungroup() %>% 
    mutate(Value_se = Value_sd/sqrt(N)) 
  
  wstat <- dat_w %>% dunn_test(AUC ~ BirthMode) %>% 
    adjust_pvalue(method = "BH") %>% 
    add_significance("p.adj") %>% 
    add_xy_position(x = "BirthMode", scales = "free", fun = "mean_sd")
    
  g <- ggplot(dat_, aes(x = BirthMode, y = Value_mean, color = BirthMode)) + 
    geom_col(aes(fill = BirthMode), color = "black") + 
    geom_errorbar(aes(ymin = Value_mean, ymax = Value_mean + Value_se),
                  width = 0.5, color = "black") + 
    geom_text(aes(y = 1, label = N), color = "gray70", size = 36/.pt, vjust = 0) + 
    stat_pvalue_manual(wstat, label = "p.adj.signif", tip.length = 0, hide.ns = T, size = 12/.pt) +
    scale_color_manual(values = BirthMode_color, aesthetics = c("color","fill")) + 
    labs(x = "", y = "AUC (mg/dL*min)", title = paste(tt, ss)) + 
    theme_classic() + theme(aspect.ratio = 1,
                            panel.background = element_rect(fill = NA),
                            plot.title = element_text(size=14,color="black",face='bold',hjust=0.5),
                            axis.title.y = element_text(size=12,color="black",face='bold'),
                            axis.text.x = element_text(size=12,color="black",face='bold'),
                            axis.text.y = element_text(size=12,color="black",face='bold'),
                            legend.text = element_text(size=12,color="black",face='bold'),
                            legend.title = element_text(size=12,color="black",face='bold'))
  g
}

g_lst <- list()
for(tt in c("GTT","ITT")){
  for(ss in c("M","F")){
    wdat_sum <- dat_sum %>% filter(Test == tt, Sex == ss)
    wdat_w <- dat_w %>% filter(Test == tt, Sex == ss)
    g_lst[[tt]][[ss]][["line"]] <- plot_time_value(wdat_sum, tt, ss)
    g_lst[[tt]][[ss]][["auc"]] <- plot_auc(wdat_w, tt, ss)
  }
}

g_lst[["GTT"]][["M"]][["line"]] <- g_lst[["GTT"]][["M"]][["line"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,850), breaks = c(0,200,400,600,800))
g_lst[["GTT"]][["F"]][["line"]] <- g_lst[["GTT"]][["F"]][["line"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,850), breaks = c(0,200,400,600,800))

g_lst[["GTT"]][["M"]][["auc"]] <- g_lst[["GTT"]][["M"]][["auc"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,4000), breaks = c(0,1000,2000,3000,4000))
g_lst[["GTT"]][["F"]][["auc"]] <- g_lst[["GTT"]][["F"]][["auc"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,4000), breaks = c(0,1000,2000,3000,4000))

p1 <- g_lst[["GTT"]][["M"]][["line"]] + g_lst[["GTT"]][["M"]][["auc"]] + 
  g_lst[["GTT"]][["F"]][["line"]] + g_lst[["GTT"]][["F"]][["auc"]] + 
  plot_layout(guides = "collect", ncol = 2) + 
  plot_annotation(title = "", tag_levels = 'a') & theme(plot.tag = element_text(size = 12, face = "bold"), 
                                                        legend.position = "right")

ggsave(filename = paste0(path_rec,"output/Glucose/GTT_panel.pdf"), device = cairo_pdf, 
       plot = p1, width = 10, height = 10, units = "in", dpi = 300)


g_lst[["ITT"]][["M"]][["line"]] <- g_lst[["ITT"]][["M"]][["line"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,260), breaks = c(0,50,100,150,200,250))
g_lst[["ITT"]][["F"]][["line"]] <- g_lst[["ITT"]][["F"]][["line"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,260), breaks = c(0,50,100,150,200,250))

g_lst[["ITT"]][["M"]][["auc"]] <- g_lst[["ITT"]][["M"]][["auc"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,700), breaks = c(0,200, 400, 600))
g_lst[["ITT"]][["F"]][["auc"]] <- g_lst[["ITT"]][["F"]][["auc"]] + 
  scale_y_continuous(expand = c(0,0), limits = c(0,700), breaks = c(0,200, 400, 600))

p2 <- g_lst[["ITT"]][["M"]][["line"]] + g_lst[["ITT"]][["M"]][["auc"]] + 
  g_lst[["ITT"]][["F"]][["line"]] + g_lst[["ITT"]][["F"]][["auc"]] + 
  plot_layout(guides = "collect", ncol = 2) + 
  plot_annotation(title = "", tag_levels = 'a') & theme(plot.tag = element_text(size = 12, face = "bold"), 
                                                        legend.position = "right")

ggsave(filename = paste0(path_rec,"output/Glucose/ITT_panel.pdf"), device = cairo_pdf, 
       plot = p2, width = 10, height = 10, units = "in", dpi = 300)


```
